refactor(studio): 拆分 Observation Inbox 领域呈现
