+++ src/studio/presentation/observation-inbox-renderer.ts
@@ -1,39 +1,24 @@
 import { DEFAULT_LANG, layout, e } from './layout.js';
 import type { Lang } from '../../shared/language.js';
-import {
-  severityReasonFor,
-  observationMetricAnnotationEntry,
-  getSkillChainAdvisory,
-  resolveAdvisoryCommand,
-  hasAssistantDeliverableArtifactText,
-  hasAssistantDeliverySignalText,
-  hasUserHardRuleText,
-  isSyntheticUserMessageText,
-  isUserInteractionMetricText,
-} from '../../observability/inbox/view-model.js';
 import type {
   ObservationInboxItem,
   ObservationInboxViewModel,
-  ObservationMetricKey,
   ResolvedObservationReviewSession,
   ResolvedOwnerSuggestion,
-  SkillChainAdvisoryCode,
-  ExperienceProblemBucket,
-  ExperienceProblemPattern,
-  ExperienceProblemSignal,
-  SkillDerivedStandard,
-  SkillLlmEnhancedReviewSections,
 } from '../../observability/inbox/view-model.js';
-import { findNegativeFeedbackMatches, findPositiveFeedbackMatches, findUserCorrectionMatches, findUserGoalShiftMatches, hasUserGoalShiftSignal } from '../../observability/inbox/feedback-projection.js';
 import { durationMsBetween } from '../../shared/time.js';
 import { incrementRecordCount } from '../../shared/record-count.js';
 import { OBSERVATION_INBOX_STYLES } from './observation-inbox/styles.js';
 import { observationInboxClientScript } from './observation-inbox/client-script.js';
+import {
+  createObservationMetricRenderers,
+  type IndicatorHelpKey,
+} from './observation-inbox/metric-renderer.js';
+import { createObservationReviewRenderers } from './observation-inbox/review-renderer.js';
+import { createObservationSignalRenderers } from './observation-inbox/signal-renderer.js';
+import { createObservationSkillChainRenderers } from './observation-inbox/skill-chain-renderer.js';
 import { createReviewerReportRenderers } from './observation-inbox/reviewer-report.js';
 import {
-  createObservationTimelineRenderers,
-  hasRepeatedExecutionSignal,
-  hasSelfCorrectionSignal,
   isAssistantCompletionResultSignal,
   isAssistantDeliverableArtifactSignal,
 } from './observation-inbox/timeline.js';
@@ -41,30 +26,18 @@ import {
   experienceSkillAnchor,
   formatTimeRange as formatTimeRangeImpl,
   formatTimestamp,
-  renderArtifactVersion,
-  renderField,
   renderJson,
   skillAnchor,
-  truncateText,
 } from './observation-inbox/helpers.js';
 import { shellQuoteArg } from '../../shared/shell-quote.js';
 export { renderFeedbackAttributionLabel } from './observation-inbox/helpers.js';
 import type {
-  ExperienceAssistiveInference,
-  ExperienceAssistiveInferenceCautionCode,
-  ExperienceAssistiveInferenceCode,
   ExperienceChecklistItem,
-  ExperienceEvidenceChain,
   ExperienceEvidenceRef,
   ExperienceFeedbackAttribution,
   ExperienceInvocation,
-  ExperienceReviewIndicators,
-  ExperienceReviewBasisCode,
   ExperienceReviewPriority,
   ExperienceReviewerReport,
-  ExperienceRuleFinding,
-  ExperienceRuleFindingCode,
-  ExperienceRuleFindingLevel,
   ExperienceEpisode,
   ExperienceFeedbackSignal,
   ExperienceSkillSegment,
@@ -174,1811 +147,67 @@ export function renderObservationInboxPage(model: ObservationInboxViewModel, lan
   const skillTimestampedInvocationCount = (skill: NonNullable<typeof experience>['skills'][number]): number =>
     skill.timestampedInvocationCount
       ?? (skill.firstSeen === '1970-01-01T00:00:00.000Z' ? 0 : skill.invocationCount);
-  const reviewSeverityMeta = (item: ObservationInboxItem): { label: string; decision: string; color: string; bg: string } => {
-    if (item.severity === 'high') {
-      return { label: '高风险/需关注', decision: '优先看，可能要补 SKILL.md 或改 skill 说明', color: 'var(--red)', bg: 'rgba(220,38,38,.08)' };
-    }
-    if (item.severity === 'medium') {
-      return { label: '低风险/抽样确认', decision: '通常不需要改 skill；抽样确认是否反复浪费时间', color: 'var(--yellow)', bg: 'rgba(202,138,4,.10)' };
-    }
-    if (item.severity === 'low') {
-      return { label: '不确定/低优先级', decision: '模型只是说不确定，不一定需要改 skill', color: 'var(--accent)', bg: 'rgba(37,99,235,.08)' };
-    }
-    return { label: '无异常/无需改 skill', decision: '更像路径、权限、文件太大或工具限制；先不当成 skill 内容缺失', color: 'var(--text-muted)', bg: 'var(--bg-muted)' };
-  };
-  const renderSeverityBadge = (item: ObservationInboxItem): string => {
-    const meta = reviewSeverityMeta(item);
-    return `<div style="display:flex;flex-direction:column;gap:3px">
-      <span style="display:inline-flex;align-items:center;width:max-content;max-width:190px;padding:3px 7px;border-radius:999px;background:${meta.bg};color:${meta.color};font-weight:650;white-space:normal">${e(meta.label)}</span>
-      <span style="color:var(--text-muted);font-size:11px">${e(meta.decision)}</span>
-    </div>`;
-  };
-  const semanticEvidence = (item: ObservationInboxItem): string => {
-    const tool = item.evidence.tool || 'Tool';
-    const target = item.evidence.query || item.evidence.path || item.evidence.assistantSnippet || '';
-    if (item.signalSubtype === 'tool_limit') return `${tool} 触发工具限制：${item.evidence.outputSnippet || target}`;
-    if (item.signalSubtype === 'transient_file_missing') return `${tool} 访问了临时文件但文件不存在：${item.evidence.path || target}`;
-    if (item.signalSubtype === 'skill_asset_read_failed') return `${tool} 读取该 skill 自身资源失败：${item.evidence.path || target}`;
-    if (item.signalSubtype === 'not_found') return `${tool} 访问了不存在的路径：${item.evidence.path || target}`;
-    if (item.signalSubtype === 'permission_denied') return `${tool} 被权限拒绝：${item.evidence.path || target}`;
-    if (item.signalSubtype === 'bash_probe') return `skill 运行过程中，agent 调用了一条 Bash 命令：${item.evidence.query || target}`;
-    if (item.signalSubtype === 'hard_miss') return `${tool} 未命中且后续未找到同主题成功证据：${target}`;
-    if (item.signalSubtype === 'exploratory_miss') return `${tool} 前序未命中，但后续有成功搜索证据：${target}`;
-    return target || item.evidence.outputSnippet || '';
-  };
-  const evidenceConclusion = (item: ObservationInboxItem): string => {
-    if (item.signalSubtype === 'bash_probe') return 'skill 运行过程中，agent 调用了一条 Bash 命令。';
-    if (item.signalSubtype === 'tool_limit') return `${item.evidence.tool || '工具'} 触发了文件太长、token 或超时限制。`;
-    if (item.signalSubtype === 'transient_file_missing') return `${item.evidence.tool || '工具'} 访问了临时文件，但文件不存在。`;
-    if (item.signalSubtype === 'skill_asset_read_failed') return `${item.evidence.tool || '工具'} 读取该 skill 自身资源失败。`;
-    if (item.signalSubtype === 'not_found') return `${item.evidence.tool || '工具'} 访问了不存在的路径。`;
-    if (item.signalSubtype === 'permission_denied') return `${item.evidence.tool || '工具'} 被权限拒绝。`;
-    if (item.signalSubtype === 'hard_miss') return `${item.evidence.tool || '工具'} 没有拿到有效结果，后续也没有看到同主题成功证据。`;
-    if (item.signalSubtype === 'exploratory_miss') return `${item.evidence.tool || '工具'} 前面没有拿到结果，但后续找到了相关内容。`;
-    if (item.signalType === 'hedging') return '模型文本里出现了不确定表达。';
-    if (item.signalType === 'explicit_marker') return '模型文本里出现了显式标记。';
-    return semanticEvidence(item);
-  };
-  const evidenceQuote = (item: ObservationInboxItem): string => {
-    return item.evidence.query || item.evidence.path || item.evidence.assistantSnippet || item.evidence.outputSnippet || '';
-  };
-  const quoteFromEvidence = (evidence: ObservationInboxItem['evidence']): string =>
-    evidence.query || evidence.path || evidence.assistantSnippet || evidence.outputSnippet || '';
-  const renderEvidenceDetailList = (item: ObservationInboxItem): string => {
-    const entries = item.representativeEvidence.length > 0 ? item.representativeEvidence : [item.evidence];
-    return `<div style="display:flex;flex-direction:column;gap:8px">
-      ${entries.map((ev, i) => {
-        const quote = quoteFromEvidence(ev);
-        const output = ev.outputSnippet && ev.outputSnippet !== quote ? ev.outputSnippet : '';
-        return `<div style="padding:9px 10px;border:1px solid var(--border);border-radius:6px;background:var(--bg-surface);text-align:left">
-          <div style="display:flex;justify-content:space-between;gap:10px;align-items:center;margin-bottom:5px">
-            <span style="font-size:12px;color:var(--text-primary);font-weight:650">Evidence ${i + 1}</span>
-            <span style="font-size:11px;color:var(--text-muted);font-family:ui-monospace,monospace">${e(ev.tool || item.signalType)}${ev.markerToken ? ` · ${e(ev.markerToken)}` : ''}</span>
-          </div>
-          <div style="font-size:12px;color:var(--text-secondary);line-height:1.45;margin-bottom:5px">${e(evidenceConclusion({ ...item, evidence: ev }))}</div>
-          ${quote ? `<pre style="margin:0;padding:8px;background:var(--bg-muted);border-radius:5px;white-space:pre-wrap;word-break:break-word;font-family:ui-monospace,monospace;font-size:11px;line-height:1.45;color:var(--text-primary);max-height:180px;overflow:auto">${e(quote)}</pre>` : ''}
-          ${output ? `<div style="margin-top:6px;color:var(--text-muted);font-size:11px">输出片段</div><pre style="margin:3px 0 0;padding:8px;background:var(--bg-muted);border-radius:5px;white-space:pre-wrap;word-break:break-word;font-family:ui-monospace,monospace;font-size:11px;line-height:1.45;color:var(--text-secondary);max-height:140px;overflow:auto">${e(output)}</pre>` : ''}
-        </div>`;
-      }).join('')}
-    </div>`;
-  };
-  const renderEvidenceCell = (item: ObservationInboxItem, max = 220): string => {
-    const quote = evidenceQuote(item);
-    return `<div style="max-width:360px">
-      <div style="font-size:12px;line-height:1.45;color:var(--text-primary)">${e(evidenceConclusion(item))}</div>
-      ${quote ? `<code style="display:block;margin-top:5px;font-family:ui-monospace,monospace;font-size:11px;line-height:1.45;color:var(--text-muted);white-space:nowrap;overflow:hidden;text-overflow:ellipsis;background:var(--bg-muted);padding:4px 6px;border-radius:4px">${e(quote.slice(0, max))}</code>` : ''}
-    </div>`;
-  };
-  const signalRuleDescription = (item: ObservationInboxItem): string => {
-    const reason = severityReasonFor(item, lang);
-    const meta = reviewSeverityMeta(item);
-    const prefix = `${meta.label}: ${item.signalType}/${item.signalSubtype}, confidence=${item.confidence.toFixed(2)}.`;
-    return `${prefix} ${reason}`;
-  };
-  const renderSignalLabel = (item: ObservationInboxItem): string => {
-    const desc = signalRuleDescription(item);
-    return `<span style="position:relative;display:inline-flex;align-items:center;gap:4px;overflow:visible">
-      <span>${e(item.signalType)}</span>
-      <span class="signal-help" tabindex="0" data-signal-title="${e(`${item.signalType}/${item.signalSubtype}`)}" data-signal-description="${e(desc)}" aria-label="${e(desc)}" style="display:inline-flex;align-items:center;justify-content:center;width:15px;height:15px;border:1px solid var(--border);border-radius:50%;font-size:10px;line-height:15px;color:var(--text-muted);cursor:help;vertical-align:top">?</span>
-    </span><br><span style="color:var(--text-muted);font-size:11px">OMK subtype: ${e(item.signalSubtype)}</span>`;
-  };
-  const renderOccurrences = (item: ObservationInboxItem): string => {
-    const text = `出现次数=${item.occurrences}。这是按 skill + cwd + signal + subtype + query/path 归一化去重后的聚合次数，不是样本量。`;
-    return `<div style="text-align:right">
-      <div style="font-weight:700">${item.occurrences}</div>
-      <div title="${e(text)}" style="color:var(--text-muted);font-size:11px;line-height:1.35;max-width:92px;margin-left:auto">同类事件次数</div>
-    </div>`;
-  };
-  const renderSourceBadge = (item: ObservationInboxItem): string => {
-    const label = item.sourceKind === 'dsh' ? 'DeepSeek Harness' : item.sourceKind === 'openclaw' ? 'OpenClaw' : item.sourceKind === 'codex' ? 'Codex' : item.sourceKind === 'markdown_log' ? 'Markdown log' : item.sourceKind === 'claude' ? 'Claude' : 'Unknown';
-    const color = item.sourceKind === 'dsh' ? '#0f766e' : item.sourceKind === 'openclaw' ? '#7c3aed' : item.sourceKind === 'codex' ? '#1677ff' : item.sourceKind === 'markdown_log' ? 'var(--green)' : item.sourceKind === 'claude' ? 'var(--accent)' : 'var(--text-muted)';
-    return `<span title="调用日志来源：${e(label)}" style="display:inline-flex;margin-top:4px;padding:2px 6px;border-radius:999px;background:var(--bg-muted);color:${color};font-size:11px;font-weight:650">${e(label)}</span>`;
-  };
-  const confidenceHeaderHelp = '判断把握：OMK 对“这条 过程发现 是否需要处理/是否高风险/需关注”的规则判断有多确定。归属把握：OMK 把这条 过程发现 归到当前 skill 名下有多确定，例如明确调用 skill 通常更高。';
-  type IndicatorHelpKey =
-    | 'userCorrection' | 'userInterruption' | 'userFollowUp' | 'negativeFeedback' | 'positiveFeedback' | 'userGoalShift' | 'hardRule' | 'selfCorrection' | 'repeatedExecution'
-    | 'toolCall' | 'toolFailure' | 'toolCancelled' | 'toolUnknown' | 'highObservation' | 'mediumObservation' | 'hedging' | 'explicitMarker'
-    | 'bash' | 'read' | 'grep' | 'bashProbe' | 'notFound' | 'toolLimit'
-    | 'skillRoleRouter' | 'skillRoleExecutor' | 'skillRoleMixed' | 'skillRoleUnknown'
-    | 'llmSkillTypeRouter' | 'llmSkillTypeDelegation' | 'llmSkillTypeExecutor' | 'llmSkillTypeAdvisory' | 'llmSkillTypeWorkflowOwner' | 'llmSkillTypeUnknown';
-  const indicatorLabels: Record<IndicatorHelpKey, string> = {
-    userCorrection: '用户纠正',
-    userInterruption: '人工中断',
-    userFollowUp: '追问/补充',
-    negativeFeedback: '负向反馈',
-    positiveFeedback: '正向反馈',
-    userGoalShift: '用户切换目标',
-    hardRule: '用户硬性要求',
-    selfCorrection: '自我纠正',
-    repeatedExecution: '重复执行',
-    toolCall: '工具调用',
-    toolFailure: '工具执行失败',
-    toolCancelled: '工具调用取消',
-    toolUnknown: '工具状态未知',
-    highObservation: '高优先级过程发现',
-    mediumObservation: '低风险/抽样过程发现',
-    hedging: '不确定表达',
-    explicitMarker: '显式缺口',
-    bash: 'Bash 调用',
-    read: 'Read 调用',
-    grep: 'Grep 调用',
-    bashProbe: 'Bash 试探',
-    notFound: '路径不存在',
-    toolLimit: '工具限制',
-    skillRoleRouter: '路由（本 session 角色）',
-    skillRoleExecutor: '执行（本 session 角色）',
-    skillRoleMixed: '路由+执行（本 session 角色）',
-    skillRoleUnknown: '未确定（本 session 角色）',
-    llmSkillTypeRouter: '路由型（能力定位）',
-    llmSkillTypeDelegation: '委派型（能力定位）',
-    llmSkillTypeExecutor: '执行型（能力定位）',
-    llmSkillTypeAdvisory: '咨询型（能力定位）',
-    llmSkillTypeWorkflowOwner: '流程负责型（能力定位）',
-    llmSkillTypeUnknown: '类型待确认（能力定位）',
-  };
-  const indicatorHelps: Record<IndicatorHelpKey, string> = {
-    userCorrection: '只统计人工用户消息里的明确纠偏表达。短词“不对 / 不是 / 错了”必须前后有空格、逗号、句号、问号等分隔；“不对称”“是不是”不会算纠正。',
-    userInterruption: '统计 Claude trace 里的 “[Request interrupted by user]” 等用户主动中断事件，表示当前执行被人工叫停。',
-    userFollowUp: '按当前报告的归因结果统计追问或补充。router / delegation 类型会把下游调用链路中的用户追问计入体验复盘；与当前 skill 无关的同窗口用户消息不计入。',
-    negativeFeedback: '统计“没用 / 垃圾 / 菜 / 做错了 / 不行 / 失败 / 看不懂 / 有问题”等负向表达。这是规则命中，不是 LLM 情绪识别。',
-    positiveFeedback: '统计“很好 / good job / 做得好 / 很棒 / 优秀 / 厉害 / 很有用 / 很有价值”等正向表达，用来保留用户认可证据。',
-    userGoalShift: '统计“换个方向 / 先不 / 不用这个 / 另一个问题”等目标切换表达。它表示当前目标可能中止或切走，不直接等同于 skill 做错。',
-    hardRule: '统计用户在对话里临时提出的硬性要求，例如“必须 / 不要 / 禁止 / 严格 / 一定要 / 只允许”。Skill 自身的强约束请看定义链路里的规则检测结果。',
-    selfCorrection: '统计 agent 在没有用户介入的情况下，发现自己的执行路径、结果或工具策略有问题并主动修正。少量说明有恢复能力，高频说明流程不稳。',
-    repeatedExecution: '统计同类步骤、工具或流程被重复执行的信号。高频出现通常对应绕路、工具策略不清晰或 workflow 缺少明确顺序。',
-    toolCall: '统计该 skill 运行片段里的 tool_use 调用总数，包括 Bash、Read、Grep 等工具。',
-    toolFailure: '统计该 skill 运行片段里失败的工具执行结果，例如 tool_result 标记 is_error=true。注意：工具执行失败不等于整个 skill 调用失败。',
-    toolCancelled: '统计 runtime 明确标记为 cancelled 的工具调用。取消与工具执行失败分开统计，也不作为知识缺口证据。',
-    toolUnknown: '统计 runtime 未提供可信终态的工具调用。状态未知不计入工具成功率或失败率分母。',
-    highObservation: '统计 severity=high 的过程发现，通常表示可能需要优先复盘的执行问题。',
-    mediumObservation: '统计 severity=medium/low 的过程发现，通常进入抽样复盘，不直接等同于必须改 skill。',
-    hedging: '统计回答或过程发现里的“不确定 / 可能 / 需要确认”等低置信文本信号。',
-    explicitMarker: '统计回答或过程发现里明确出现“未知 / 知识缺口 / 没覆盖”等显式标记。',
-    bash: '统计该 skill 运行期间调用 Bash 工具的次数。',
-    read: '统计该 skill 运行期间调用 Read 工具的次数。',
-    grep: '统计该 skill 运行期间调用 Grep 工具的次数。',
-    bashProbe: '统计过程发现中被判断为 Bash 试探的次数，例如命令看起来是在试目录、试路径或探测环境。',
-    notFound: '统计过程发现中路径或文件不存在的次数。',
-    toolLimit: '统计文件太长、token 上限、超时等工具限制导致的问题。',
-    skillRoleRouter: '本 session 内 skill 主要在调用 Skill 工具触发下游能力，自己不直接产出。判定来自 trace 行为，看到 Skill tool_use 调用占主导。常见于调度型 skill。',
-    skillRoleExecutor: '本 session 内 skill 主要在调用具体工具（Bash / Read / Write / Edit）直接产出 artifact。判定来自 trace 行为。常见于编码 / 文档生成 / 数据处理类 skill。',
-    skillRoleMixed: '本 session 内 skill 既调用 Skill 工具触发下游，又自己调用执行工具直接产出。trace 行为同时看到两种特征。常见于负责调度但保留兜底执行的 skill。',
-    skillRoleUnknown: '本 session 内 trace 行为没显示明显的路由或执行特征。可能是 advisory / 对话型 skill，或样本太少没足够证据。',
-    llmSkillTypeRouter: 'LLM 综合 SKILL.md 描述和 trace 行为判定为路由型：定位是把请求分发到下游，自身不直接产出。和"路由（本 session 角色）"区别：这个是 skill 本质定位，不只是单次 session 表现。',
-    llmSkillTypeDelegation: 'LLM 判定为委派型：把任务交给 child session / 子 agent 执行，自己只负责拆任务、监督生命周期和回收结果。区别于路由型在于：路由型主要做转发，委派型还要管 child 跑偏 / 启动 / 终止全生命周期。',
-    llmSkillTypeExecutor: 'LLM 判定为执行型：定位是自己直接干活产出 artifact。和"执行（本 session 角色）"区别：这个是 skill 本质定位，不只是单次 session 表现。',
-    llmSkillTypeAdvisory: 'LLM 判定为咨询型：定位是回答问题 + 给证据，不一定产出 artifact。常见于审计 / 分析 / 解释类 skill。本 session 角色一般会显示"未确定"，因为 advisory 没有明显工具使用特征。',
-    llmSkillTypeWorkflowOwner: 'LLM 判定为流程负责型：定位是管理一条多阶段 workflow 的闭环状态。它可以委派其他 skill 执行阶段动作，但要负责阶段矩阵、产物、异常和用户反馈闭环。',
-    llmSkillTypeUnknown: 'LLM 没能从 SKILL.md 和 trace 中得到足够证据判定 skill 类型。可能是 SKILL.md 描述不清，或 trace 样本不足。建议 skill owner 补充 SKILL.md 顶部声明，或多观察几个 session。',
-  };
-  const metric = (label: string, value: number, helpKey: IndicatorHelpKey, title?: string): string =>
-    `<button type="button" class="metric-item" data-metric-key="${helpKey}" onclick="openMetricGuide('${helpKey}')" title="${e(title ?? `点击查看“${label}”指标说明`)}"><span>${e(label)}</span> <strong>${value}</strong></button>`;
-  const formatEntrypoint = (value?: string): string => {
-    const labels: Record<string, string> = {
-      'claude-desktop': 'Claude Code App',
-      cli: 'Claude CLI',
-      'sdk-ts': 'Claude SDK',
-      'codex-desktop': 'ChatGPT 桌面端（Codex）',
-      'codex-cli': 'Codex CLI',
-      'codex-sdk': 'Codex SDK',
-      'codex-vscode': 'Codex VS Code',
-      claudian: 'Claudian',
-      openclaw: 'OpenClaw',
-      markdown_log: 'Markdown 日志',
-    };
-    return value ? (labels[value] ?? value) : '未记录';
-  };
-  const formatAttributionSource = (value?: string): string => {
-    if (value === 'command-name') return '通过斜杠命令触发';
-    if (value === 'business-action' || value === legacyBusinessActionSource()) return '通过业务动作触发';
-    if (value === 'skill-tool') return '通过 Skill 工具启动';
-    if (value === 'read-skill-md') return '通过读取 Skill 文档推断';
-    if (value === 'skill-script') return '通过 Skill 脚本路径推断';
-    if (value === 'general') return '未识别到具体 skill';
-    return value || '未记录';
-  };
-  const renderInvocationSummary = (
-    indicators: {
-      userInterruptionCount: number;
-      toolFailureCount: number;
-      toolCancelledCount?: number;
-      toolUnknownCount?: number;
-      toolCallCount: number;
-    },
-    invocationCount: number,
-  ): string => {
-    const total = indicators.toolCallCount;
-    const failed = Math.max(0, indicators.toolFailureCount);
-    const cancelled = Math.max(0, indicators.toolCancelledCount ?? 0);
-    const unknown = Math.max(0, indicators.toolUnknownCount ?? 0);
-    const interrupted = Math.max(0, indicators.userInterruptionCount);
-    const success = Math.max(0, total - failed - cancelled - unknown);
-    const pct = (value: number): string => total > 0 ? `${Math.round(value / total * 100)}%` : '—';
-    return `<div class="invocation-summary" title="这是 trace 中可观测到的调用过程汇总。工具执行失败/人工中断是过程信号，不直接等同于整个 skill 失败或用户目标失败。">
-      <div class="invocation-total">工具调用总次数 <strong>${total}</strong></div>
-      <div class="invocation-breakdown">
-        <span>工具执行成功 ${success} / ${pct(success)}</span>
-        <span>工具执行失败 ${failed} / ${pct(failed)}</span>
-        ${cancelled > 0 ? `<span>工具调用取消 ${cancelled} / ${pct(cancelled)}</span>` : ''}
-        ${unknown > 0 ? `<span>结果状态未知 ${unknown} / ${pct(unknown)}</span>` : ''}
-        <span>人工中断 ${interrupted} 次</span>
-      </div>
-      <div class="invocation-footnote">Skill 调用段 ${invocationCount}</div>
-    </div>`;
-  };
-  const pct = (value: number, total: number): string => total > 0 ? `${Math.round(value / total * 100)}%` : '—';
-  const EMPTY_VALUE = '—';
-  const renderShare = (count: number, total: number): string => {
-    if (count <= 0 && total <= 0) {
-      return `<span class="summary-count">${EMPTY_VALUE}</span>`;
-    }
-    const safeTotal = Math.max(total, count);
-    return `<span class="summary-count">${count}</span><span class="summary-pct">/ ${pct(count, safeTotal)}</span>`;
-  };
-  const renderMetric = (label: string, count: number, unit = ''): string => {
-    const empty = count <= 0;
-    return `<span class="summary-metric${empty ? ' summary-metric-empty' : ''}"><span class="summary-name">${e(label)}</span><span class="summary-count">${empty ? EMPTY_VALUE : count}</span>${unit && !empty ? `<span class="summary-unit-text">${e(unit)}</span>` : ''}</span>`;
-  };
-  const renderMetricShare = (label: string, count: number, total: number): string => {
-    const empty = count <= 0;
-    return `<span class="summary-metric${empty ? ' summary-metric-empty' : ''}"><span class="summary-name">${e(label)}</span>${renderShare(count, total)}</span>`;
-  };
-  const reviewImpactTitle = (level: 'priority' | 'sample'): string =>
-    level === 'priority'
-      ? '命中后会影响“优先复盘”：建议先打开上下文看证据。'
-      : '命中后会影响“抽样复盘”：建议抽样打开上下文确认。';
-  const renderDecisionMetric = (label: string, count: number, level: 'priority' | 'sample', unit = ''): string => {
-    const active = count > 0;
-    return `<span class="summary-metric${active ? ` summary-impact summary-impact-${level}` : ' summary-metric-empty'}"${active ? ` title="${e(reviewImpactTitle(level))}"` : ''}><span class="summary-name">${e(label)}</span><span class="summary-count">${active ? count : EMPTY_VALUE}</span>${unit && active ? `<span class="summary-unit-text">${e(unit)}</span>` : ''}</span>`;
-  };
-  const renderDecisionMetricIfPositive = (label: string, count: number, level: 'priority' | 'sample', unit = ''): string =>
-    count > 0 ? renderDecisionMetric(label, count, level, unit) : '';
-  const renderDecisionMetricShare = (label: string, count: number, total: number, level: 'priority' | 'sample'): string => {
-    const active = count > 0;
-    return `<span class="summary-metric${active ? ` summary-impact summary-impact-${level}` : ' summary-metric-empty'}"${active ? ` title="${e(reviewImpactTitle(level))}"` : ''}><span class="summary-name">${e(label)}</span>${renderShare(count, total)}</span>`;
-  };
-  // 结果类观察指标：>0 时显示轻微 tag（complete chip 视觉），
-  // =0 时与其他指标一样走 summary-metric-empty。不接入复盘评分。
-  const renderSoftMetric = (label: string, count: number, title?: string): string => {
-    const active = count > 0;
-    return `<span class="summary-metric${active ? ' summary-impact summary-impact-soft' : ' summary-metric-empty'}"${active && title ? ` title="${e(title)}"` : ''}><span class="summary-name">${e(label)}</span><span class="summary-count">${active ? count : EMPTY_VALUE}</span></span>`;
-  };
-  const renderRankedCounts = (
-    counts: Record<string, number> | undefined,
-    total: number,
-    options: { max?: number; label?: (key: string) => string } = {},
-  ): string => {
-    const entries = Object.entries(counts ?? {})
-      .filter(([, count]) => count > 0)
-      .sort((a, b) => b[1] - a[1])
-      .slice(0, options.max ?? 4);
-    if (entries.length === 0) return '<span class="summary-muted">无明细</span>';
-    return entries.map(([key, count]) => renderMetricShare(options.label ? options.label(key) : key, count, total)).join('<span class="summary-sep">，</span>');
-  };
-  const userFacingToolLabel = (tool: string): string => {
-    const lower = tool.toLowerCase();
-    if (['exec', 'process', 'bash', 'shell', 'run'].includes(lower)) return '命令执行';
-    if (lower === 'read') return '读取文件';
-    if (['grep', 'glob', 'search', 'find'].includes(lower)) return '搜索文件';
-    if (['edit', 'write', 'multiedit'].includes(lower)) return '修改文件';
-    if (lower === 'todowrite') return '记录待办';
-    if (lower === 'skill') return '启动 skill';
-    return tool;
-  };
-  const mergeCountsByLabel = (counts: Record<string, number> | undefined, label: (key: string) => string): Record<string, number> => {
-    const next: Record<string, number> = {};
-    for (const [key, count] of Object.entries(counts ?? {})) {
-      const display = label(key);
-      incrementRecordCount(next, display, count);
-    }
-    return next;
-  };
-  const compactRankedCountText = (
-    counts: Record<string, number> | undefined,
-    options: { max?: number; label?: (key: string) => string } = {},
-  ): string => {
-    const entries = Object.entries(counts ?? {})
-      .filter(([, count]) => count > 0)
-      .sort((a, b) => b[1] - a[1])
-      .slice(0, options.max ?? 2);
-    if (entries.length === 0) return '未记录';
-    return entries.map(([key, count]) => `${options.label ? options.label(key) : key} ${count}`).join('、');
-  };
-  const sourceMetadataLine = (label: string, counts: Record<string, number> | undefined, options: { max?: number; valueLabel?: (value: string) => string } = {}): string => {
-    const text = compactRankedCountText(counts, { max: options.max ?? 2, label: options.valueLabel });
-    return text === '未记录' ? '' : `<div>${e(label)}：${e(text)}</div>`;
-  };
-  const renderOpenClawSourceMetadata = (counts: NonNullable<NonNullable<typeof experience>['skills'][number]['sourceMetadataCounts']> | undefined): string => {
-    if (!counts) return '';
-    const lines = [
-      sourceMetadataLine('渠道', counts.channels, { valueLabel: formatChannel }),
-      sourceMetadataLine('用户', counts.senders, { max: 1 }),
-      sourceMetadataLine('业务动作', sourceMetadataBusinessActionCounts(counts)),
-      sourceMetadataLine('模型', counts.models),
-      sourceMetadataLine('供应商', counts.providers),
-    ].filter(Boolean);
-    return lines.length > 0 ? `<div class="openclaw-source-meta">${lines.join('')}</div>` : '';
-  };
-  const renderSessionOpenClawSourceMetadata = (session: ExperienceSessionSummary): string => {
-    const meta = session.sourceMetadata;
-    if (!meta) return '';
-    const lines = [
-      meta.channel ? `渠道：${formatChannel(meta.channel)}` : '',
-      meta.sender || meta.senderId ? `用户：${meta.sender ?? ''}${meta.senderId ? `(${meta.senderId})` : ''}` : '',
-      sourceMetadataBusinessActions(meta).length ? `业务动作：${sourceMetadataBusinessActions(meta).join('、')}` : '',
-      meta.model ? `模型：${meta.model}` : '',
-      meta.provider ? `供应商：${meta.provider}` : '',
-    ].filter(Boolean);
-    return lines.length > 0 ? `<span style="font-family:system-ui,-apple-system,BlinkMacSystemFont,Segoe UI,sans-serif;color:var(--text-muted)">${e(lines.join(' · '))}</span>` : '';
-  };
-  const formatChannel = (value: string): string => {
-    const labels: Record<string, string> = {
-      [legacyBusinessChannel()]: '业务通道',
-      dingtalk: '钉钉',
-      cli: 'CLI',
-    };
-    return labels[value] ?? value;
-  };
-  const legacyBusinessActionSource = (): string => ['ai', 'ma-cmd'].join('');
-  const legacyBusinessActionField = (): string => ['ai', 'maCommands'].join('');
-  const legacyBusinessChannel = (): string => ['ai', 'ma'].join('');
-  const sourceMetadataBusinessActions = (meta: ExperienceSessionSummary['sourceMetadata']): string[] => {
-    if (!meta) return [];
-    const legacy = (meta as unknown as Record<string, unknown>)[legacyBusinessActionField()];
-    const legacyActions = Array.isArray(legacy) ? legacy.filter((item): item is string => typeof item === 'string') : [];
-    return Array.from(new Set([...(meta.businessActions ?? []), ...legacyActions])).sort();
-  };
-  const sourceMetadataBusinessActionCounts = (
-    counts: NonNullable<NonNullable<typeof experience>['skills'][number]['sourceMetadataCounts']> | undefined,
-  ): Record<string, number> | undefined => {
-    if (!counts) return undefined;
-    const legacy = (counts as unknown as Record<string, unknown>)[legacyBusinessActionField()];
-    const merged = { ...(counts.businessActions ?? {}) };
-    if (legacy && typeof legacy === 'object' && !Array.isArray(legacy)) {
-      for (const [key, value] of Object.entries(legacy as Record<string, unknown>)) {
-        if (typeof value === 'number') incrementRecordCount(merged, key, value);
-      }
-    }
-    return merged;
-  };
-  const problemBucketLabels: Record<ExperienceProblemBucket, string> = {
-    output_format: '输出格式不符合预期',
-    content_accuracy: '内容理解可能不对',
-    missing_context: '可能没读到关键资料',
-    rule_violation: '用户明确提出要求',
-    workflow_mismatch: '可能走错流程',
-    tool_runtime: '工具/环境问题',
-    goal_shift: '用户中途改了目标',
-    unclear: '问题类型不明确',
-  };
-  const problemSignalLabels: Record<ExperienceProblemSignal, string> = {
-    user_correction: '用户纠正',
-    negative_feedback: '负向反馈',
-    user_interruption: '人工中断',
-    hard_rule: '用户硬性要求',
-    user_goal_shift: '目标变化',
-    tool_failure: '工具失败',
-    workflow_mismatch: '流程不匹配',
-    artifact_missing: '产物缺失',
-    observer_lifecycle_failed: '观察器异常',
-    orchestration_boundary_violation: '调度边界异常',
-  };
-  const timelineTagForProblemPattern = (pattern: ExperienceProblemPattern): string => {
-    if (pattern.signalTypes.includes('user_correction')) return 'user_correction';
-    if (pattern.signalTypes.includes('negative_feedback')) return 'negative_feedback';
-    if (pattern.signalTypes.includes('user_interruption')) return 'user_interruption';
-    if (pattern.signalTypes.includes('hard_rule')) return 'hard_rule';
-    if (pattern.signalTypes.includes('user_goal_shift')) return 'user_goal_shift';
-    if (pattern.signalTypes.includes('tool_failure')) return 'tool_failure';
-    return '';
-  };
-  const patternKeywordLabel = (pattern: ExperienceProblemPattern): string => {
-    const raw = pattern.patternKey.split(':').slice(1).join(':');
-    const labels: Record<string, string> = {
-      prd: 'PRD',
-      demo: 'Demo',
-      format: '格式/模板',
-      figma: '设计稿',
-      context: '上下文/资料',
-      schema: '字段/接口/路径',
-      workflow: '流程',
-      rule: '规则',
-      wrong: '不符合预期',
-      tool_limit: '工具限制',
-      tool_failure: '工具失败',
-      not_found: '文件/路径不存在',
-      permission: '权限问题',
-      timeout: '超时',
-    };
-    const parts = raw.split('+').map((part) => labels[part] ?? part).filter(Boolean);
-    return parts.length > 0 ? parts.slice(0, 3).join(' + ') : problemBucketLabels[pattern.bucket];
-  };
-  const renderSkillProblemPatterns = (skillName: string, patterns: ExperienceProblemPattern[] | undefined): string => {
-    const top = (patterns ?? []).slice(0, 3);
-    if (top.length === 0) return '';
-    return `<div class="summary-row problem-pattern-row"><span class="summary-title">【发现问题线索】</span><span class="problem-pattern-list">${top.map((pattern) => {
-      const rawSessionId = pattern.recentSessionIds[0] ?? '';
-      const sessionId = experienceSessionIdBySkillAndSession.get(`${skillName}\u0000${rawSessionId}`) ?? '';
-      const tag = timelineTagForProblemPattern(pattern);
-      const signals = pattern.signalTypes.map((signal) => problemSignalLabels[signal]).join('、');
-      const title = `${problemBucketLabels[pattern.bucket]}：${patternKeywordLabel(pattern)}。命中 ${pattern.count} 次，涉及 ${pattern.sessionCount} 个 session。来源：${signals || '规则聚合'}。`;
-      return `<button type="button" class="problem-pattern-chip" data-no-rollup-click="1"${sessionId ? ` data-open-experience-session="${e(sessionId)}"` : ''}${tag ? ` data-open-timeline-tag="${e(tag)}"` : ''} title="${e(title)}"><span class="pattern-bucket">${e(problemBucketLabels[pattern.bucket])}</span><span class="pattern-key">${e(patternKeywordLabel(pattern))}</span><span class="pattern-count">${pattern.count} 次 / ${pattern.sessionCount} 个 session</span></button>`;
-    }).join('')}</span></div>`;
-  };
-  const ZERO_DISPLAY_INDICATORS: ExperienceReviewIndicators = {
-    userMessageCount: 0,
-    userFollowUpCount: 0,
-    userCorrectionCount: 0,
-    userInterruptionCount: 0,
-    sessionInterruptedCount: 0,
-    negativeFeedbackCount: 0,
-    positiveFeedbackCount: 0,
-    userGoalShiftCount: 0,
-    hardRuleTextHitCount: 0,
-	    assistantDeliverySignalCount: 0,
-	    deliverableArtifactSignalCount: 0,
-	    routerDownstreamCompleted: 0,
-	    routerDownstreamFailed: 0,
-	    selfCorrectionCount: 0,
-    repeatedExecutionCount: 0,
-    toolCallCount: 0,
-    toolFailureCount: 0,
-    toolCancelledCount: 0,
-    toolUnknownCount: 0,
-    highObservationCount: 0,
-    mediumObservationCount: 0,
-    hedgingCount: 0,
-    explicitMarkerCount: 0,
-  };
-  const USER_INTERRUPTION_DISPLAY_RE = /\[Request interrupted by user(?: for tool use)?\]|interrupted by user|用户中断/i;
-  const displayRegexMatches = (pattern: RegExp, value: string): boolean => {
-    pattern.lastIndex = 0;
-    return pattern.test(value);
-  };
-  const displayMetricVerdict = (event: ExperienceTimelineEvent, metricKey: ObservationMetricKey, metricScopeId?: string): 'confirmed' | 'rejected' | '' => {
-    const entry = observationMetricAnnotationEntry(
-      reviewState,
-      { ...event, metricScopeId },
-      metricKey,
-    );
-    return entry?.verdict === 'confirmed' || entry?.verdict === 'rejected' ? entry.verdict : '';
-  };
-  const displayMetricIsActive = (event: ExperienceTimelineEvent, metricKey: ObservationMetricKey, ruleDetected: boolean, metricScopeId?: string): boolean => {
-    const verdict = displayMetricVerdict(event, metricKey, metricScopeId);
-    if (verdict === 'confirmed') return true;
-    if (verdict === 'rejected') return false;
-    return ruleDetected;
-  };
-  const displayMetricCount = (event: ExperienceTimelineEvent, metricKey: ObservationMetricKey, ruleCount: number, metricScopeId?: string): number => {
-    const verdict = displayMetricVerdict(event, metricKey, metricScopeId);
-    if (verdict === 'confirmed') return Math.max(1, ruleCount);
-    if (verdict === 'rejected') return 0;
-    return ruleCount;
-  };
-  const shouldIncludeDownstreamFeedbackForDisplay = (session: ExperienceSessionSummary): boolean => {
-    const episodes = session.sessionStory?.episodes ?? [];
-    const segmentIds = new Set(episodes.flatMap((episode) =>
-      (episode.skillSegments ?? [])
-        .filter((segment) => segment.skillName === session.skillName)
-        .map((segment) => segment.id)
-    ));
-    if (segmentIds.size === 0) return false;
-    const ownSegments = episodes.flatMap((episode) => episode.skillSegments ?? [])
-      .filter((segment) => segment.skillName === session.skillName);
-    if (ownSegments.some((segment) =>
-      segment.skillType === 'router'
-      || segment.skillType === 'delegation'
-      || segment.episodeRole === 'router'
-      || segment.episodeRole === 'delegator'
-    )) return true;
-    return episodes.some((episode) =>
-      (episode.orchestrationEdges ?? []).some((edge) =>
-        Boolean(edge.parentSkillSegmentId && segmentIds.has(edge.parentSkillSegmentId))
-      )
-    );
-  };
-  const feedbackAttributionBelongsToDisplaySession = (
-    attribution: ExperienceFeedbackAttribution,
-    session: ExperienceSessionSummary,
-    includeDownstream = shouldIncludeDownstreamFeedbackForDisplay(session),
-  ): boolean =>
-    attribution.skillName === session.skillName
-    && (
-      attribution.attributionRole === 'primary_fault'
-      || includeDownstream && attribution.attributionRole === 'downstream_related'
-    );
-  const canonicalFeedbackSignalsForDisplay = (
-    session: ExperienceSessionSummary,
-    signalType?: ExperienceFeedbackSignal['type'],
-  ): ExperienceFeedbackSignal[] => {
-    const includeDownstream = shouldIncludeDownstreamFeedbackForDisplay(session);
-    return (session.sessionStory?.episodes?.flatMap((episode) => episode.feedbackSignals ?? []) ?? [])
-      .filter((signal) =>
-        (!signalType || signal.type === signalType)
-        && displayFeedbackSignalIsActive(signal, session)
-        && (signal.canonicalAttributions ?? signal.attributions ?? []).some((attribution) =>
-          feedbackAttributionBelongsToDisplaySession(attribution, session, includeDownstream)
-        )
-      );
-  };
-  const metricKeyForFeedbackSignal = (signal: ExperienceFeedbackSignal): ObservationMetricKey | undefined => {
-    if (signal.type === 'follow_up') return 'user_follow_up';
-    if (signal.type === 'correction') return 'user_correction';
-    if (signal.type === 'interruption') return 'user_interruption';
-    if (signal.type === 'frustration') return 'negative_feedback';
-    if (signal.type === 'positive') return 'positive_feedback';
-    return undefined;
-  };
-  const displayFeedbackSignalIsActive = (signal: ExperienceFeedbackSignal, session: ExperienceSessionSummary): boolean => {
-    const metricKey = metricKeyForFeedbackSignal(signal);
-    if (!metricKey) return true;
-    const verdict = displayMetricVerdict(signal.evidenceRef as ExperienceTimelineEvent, metricKey, session.id);
-    if (verdict === 'confirmed') return true;
-    if (verdict === 'rejected') return false;
-    return true;
-  };
-  const feedbackSignalTypeForMetric = (metricKey: ObservationMetricKey): ExperienceFeedbackSignal['type'] | undefined => {
-    if (metricKey === 'user_follow_up') return 'follow_up';
-    if (metricKey === 'user_correction') return 'correction';
-    if (metricKey === 'user_interruption') return 'interruption';
-    if (metricKey === 'negative_feedback') return 'frustration';
-    if (metricKey === 'positive_feedback') return 'positive';
-    return undefined;
-  };
-  const canonicalFeedbackCountsForDisplay = (session: ExperienceSessionSummary): Pick<ExperienceReviewIndicators, 'userFollowUpCount' | 'userCorrectionCount' | 'userInterruptionCount' | 'negativeFeedbackCount' | 'positiveFeedbackCount'> | undefined => {
-    const signals = session.sessionStory?.episodes?.flatMap((episode) => episode.feedbackSignals ?? []) ?? [];
-    if (signals.length === 0) return undefined;
-    const owned = canonicalFeedbackSignalsForDisplay(session);
-    return {
-      userFollowUpCount: owned.filter((signal) => signal.type === 'follow_up').length,
-      userCorrectionCount: owned.filter((signal) => signal.type === 'correction').length,
-      userInterruptionCount: owned.filter((signal) => signal.type === 'interruption').length,
-      negativeFeedbackCount: owned.filter((signal) => signal.type === 'frustration').length,
-      positiveFeedbackCount: owned.filter((signal) => signal.type === 'positive').length,
-    };
-  };
-  const displayIndicatorsForSession = (session: ExperienceSessionSummary): ExperienceReviewIndicators => {
-    const humanEvents = (session.timelinePreview ?? []).filter((event) => event.kind === 'user_message' && Boolean(event.snippet) && !isSyntheticUserMessageText(event.snippet ?? ''));
-    const interactionEvents = humanEvents.filter((event) => isUserInteractionMetricText(event.snippet ?? ''));
-    const canonicalFeedback = canonicalFeedbackCountsForDisplay(session);
-    return {
-      ...session.indicators,
-      userFollowUpCount: canonicalFeedback?.userFollowUpCount ?? interactionEvents.reduce((sum, event, index) => sum + (displayMetricIsActive(event, 'user_follow_up', index > 0 && !hasUserGoalShiftSignal(event.snippet ?? ''), session.id) ? 1 : 0), 0),
-      userCorrectionCount: canonicalFeedback?.userCorrectionCount ?? interactionEvents.reduce((sum, event) => sum + displayMetricCount(event, 'user_correction', findUserCorrectionMatches(event.snippet ?? '').length, session.id), 0),
-      userInterruptionCount: canonicalFeedback?.userInterruptionCount ?? interactionEvents.reduce((sum, event) => sum + (displayMetricIsActive(event, 'user_interruption', displayRegexMatches(USER_INTERRUPTION_DISPLAY_RE, event.snippet ?? ''), session.id) ? 1 : 0), 0),
-      negativeFeedbackCount: canonicalFeedback?.negativeFeedbackCount ?? interactionEvents.reduce((sum, event) => sum + displayMetricCount(event, 'negative_feedback', findNegativeFeedbackMatches(event.snippet ?? '').length), 0),
-      positiveFeedbackCount: canonicalFeedback?.positiveFeedbackCount ?? interactionEvents.reduce((sum, event) => sum + displayMetricCount(event, 'positive_feedback', findPositiveFeedbackMatches(event.snippet ?? '').length), 0),
-      userGoalShiftCount: interactionEvents.reduce((sum, event) => sum + displayMetricCount(event, 'user_goal_shift', findUserGoalShiftMatches(event.snippet ?? '').length, session.id), 0),
-      hardRuleTextHitCount: interactionEvents.reduce((sum, event) => sum + (displayMetricIsActive(event, 'hard_rule', hasUserHardRuleText(event.snippet ?? ''), session.id) ? 1 : 0), 0),
-      assistantDeliverySignalCount: (session.timelinePreview ?? []).reduce((sum, event) => sum + (displayMetricIsActive(event, 'completion_result', event.kind === 'assistant_message' && hasAssistantDeliverySignalText(event.fullText ?? event.snippet ?? '')) ? 1 : 0), 0),
-      deliverableArtifactSignalCount: (session.timelinePreview ?? []).reduce((sum, event) => sum + (displayMetricIsActive(event, 'deliverable_artifact', event.kind === 'assistant_message' && hasAssistantDeliverableArtifactText(event.fullText ?? event.snippet ?? '')) ? 1 : 0), 0),
-      selfCorrectionCount: (session.timelinePreview ?? []).reduce((sum, event) => sum + (displayMetricIsActive(event, 'self_correction', hasSelfCorrectionSignal(event)) ? 1 : 0), 0),
-      repeatedExecutionCount: (session.timelinePreview ?? []).reduce((sum, event) => sum + (displayMetricIsActive(event, 'repeated_execution', hasRepeatedExecutionSignal(event)) ? 1 : 0), 0),
-    };
-  };
-  const sessionMetricSourceTitle = (session: ExperienceSessionSummary, metricKey: ObservationMetricKey, label: string): string => {
-    const canonicalType = feedbackSignalTypeForMetric(metricKey);
-    if (canonicalType) {
-      const signals = canonicalFeedbackSignalsForDisplay(session, canonicalType);
-      if (signals.length > 0) {
-        const rows = signals.slice(0, 8).map((signal) => {
-          const role = (signal.canonicalAttributions ?? signal.attributions ?? [])
-            .filter((attribution) => feedbackAttributionBelongsToDisplaySession(attribution, session))
-            .map((attribution) => attribution.attributionRole === 'downstream_related' ? '下游调用链路' : '直接归因')
-            .find(Boolean) ?? '归因命中';
-          return `#${signal.evidenceRef.messageIndex ?? '—'} ${role}：${signal.text.slice(0, 80).replace(/\s+/g, ' ')}`;
-        });
-        return [
-          `${label}：最终计入 ${signals.length}`,
-          '来源：任务执行归因模型',
-          `证据：${rows.join('；')}`,
-          '点击原文回溯中的同名标签，可以定位到这些 evidenceRef。',
-        ].join('\n');
-      }
-    }
-    const humanEvents = (session.timelinePreview ?? []).filter((event) => event.kind === 'user_message' && Boolean(event.snippet) && !isSyntheticUserMessageText(event.snippet ?? ''));
-    const rows: string[] = [];
-    let ruleCount = 0;
-    let confirmed = 0;
-    let rejected = 0;
-    let finalCount = 0;
-    let interactionIndex = 0;
-    humanEvents.forEach((event) => {
-      const text = event.snippet ?? '';
-      if (!isUserInteractionMetricText(text)) return;
-      const effectiveIndex = interactionIndex;
-      interactionIndex += 1;
-      const rule = metricKey === 'user_follow_up'
-        ? (effectiveIndex > 0 && !hasUserGoalShiftSignal(text) ? 1 : 0)
-        : metricKey === 'user_correction'
-          ? findUserCorrectionMatches(text).length
-          : metricKey === 'user_interruption'
-            ? (displayRegexMatches(USER_INTERRUPTION_DISPLAY_RE, text) ? 1 : 0)
-            : metricKey === 'negative_feedback'
-              ? findNegativeFeedbackMatches(text).length
-              : metricKey === 'positive_feedback'
-                ? findPositiveFeedbackMatches(text).length
-                : metricKey === 'hard_rule'
-                  ? (hasUserHardRuleText(text) ? 1 : 0)
-                  : metricKey === 'user_goal_shift'
-                    ? findUserGoalShiftMatches(text).length
-                    : 0;
-      const verdict = displayMetricVerdict(event, metricKey, session.id);
-      const finalValue = displayMetricCount(event, metricKey, rule, session.id);
-      ruleCount += rule;
-      if (verdict === 'confirmed') confirmed += 1;
-      if (verdict === 'rejected') rejected += 1;
-      finalCount += finalValue;
-      if (rule > 0 || verdict) {
-        rows.push(`#${event.messageIndex ?? '—'} ${verdict === 'confirmed' ? '人工同意' : verdict === 'rejected' ? '人工反对' : '规则命中'}：${text.slice(0, 80).replace(/\s+/g, ' ')}`);
-      }
-    });
-    return [
-      `${label}：最终计入 ${finalCount}`,
-      `规则命中 ${ruleCount}，人工同意 ${confirmed}，人工反对 ${rejected}`,
-      rows.length > 0 ? `来源：${rows.join('；')}` : '来源：无',
-      '点击按钮查看指标定义；人工反对只扣掉对应消息，不会自动扣掉其它来源。',
-    ].join('\n');
-  };
-  const displayIndicatorsBySkill = new Map<string, ExperienceReviewIndicators>();
-  for (const session of experience?.sessions ?? []) {
-    const current = displayIndicatorsBySkill.get(session.skillName) ?? { ...ZERO_DISPLAY_INDICATORS };
-    const next = displayIndicatorsForSession(session);
-    displayIndicatorsBySkill.set(session.skillName, {
-      userMessageCount: current.userMessageCount + next.userMessageCount,
-      userFollowUpCount: current.userFollowUpCount + next.userFollowUpCount,
-      userCorrectionCount: current.userCorrectionCount + next.userCorrectionCount,
-      userInterruptionCount: current.userInterruptionCount + next.userInterruptionCount,
-      sessionInterruptedCount: current.sessionInterruptedCount + (next.sessionInterruptedCount ?? 0),
-      negativeFeedbackCount: current.negativeFeedbackCount + next.negativeFeedbackCount,
-      positiveFeedbackCount: current.positiveFeedbackCount + next.positiveFeedbackCount,
-      userGoalShiftCount: current.userGoalShiftCount + next.userGoalShiftCount,
-      hardRuleTextHitCount: current.hardRuleTextHitCount + next.hardRuleTextHitCount,
-	      assistantDeliverySignalCount: current.assistantDeliverySignalCount + (next.assistantDeliverySignalCount ?? 0),
-	      deliverableArtifactSignalCount: current.deliverableArtifactSignalCount + (next.deliverableArtifactSignalCount ?? 0),
-	      routerDownstreamCompleted: current.routerDownstreamCompleted + (next.routerDownstreamCompleted ?? 0),
-	      routerDownstreamFailed: current.routerDownstreamFailed + (next.routerDownstreamFailed ?? 0),
-	      selfCorrectionCount: current.selfCorrectionCount + (next.selfCorrectionCount ?? 0),
-      repeatedExecutionCount: current.repeatedExecutionCount + (next.repeatedExecutionCount ?? 0),
-      toolCallCount: current.toolCallCount + next.toolCallCount,
-      toolFailureCount: current.toolFailureCount + next.toolFailureCount,
-      toolCancelledCount: (current.toolCancelledCount ?? 0) + (next.toolCancelledCount ?? 0),
-      toolUnknownCount: (current.toolUnknownCount ?? 0) + (next.toolUnknownCount ?? 0),
-      highObservationCount: current.highObservationCount + next.highObservationCount,
-      mediumObservationCount: current.mediumObservationCount + next.mediumObservationCount,
-      hedgingCount: current.hedgingCount + next.hedgingCount,
-      explicitMarkerCount: current.explicitMarkerCount + next.explicitMarkerCount,
-    });
-  }
-  const displayBasisCodes = (indicators: ExperienceReviewIndicators): ExperienceReviewBasisCode[] => {
-    const codes: ExperienceReviewBasisCode[] = [];
-    if (indicators.highObservationCount > 0) codes.push('has_high_observation');
-    if (indicators.mediumObservationCount > 0) codes.push('has_medium_observation');
-    if (indicators.userCorrectionCount > 0) codes.push('user_correction');
-    if (indicators.userInterruptionCount > 0) codes.push('user_interruption');
-    if (indicators.sessionInterruptedCount > 0) codes.push('session_interrupted');
-    if (indicators.negativeFeedbackCount > 0) codes.push('negative_feedback');
-    if (indicators.hardRuleTextHitCount > 0) codes.push('hard_rule_text_hit');
-    if (indicators.toolFailureCount > 0) codes.push('tool_failure');
-    if (indicators.hedgingCount > 0) codes.push('hedging_signal');
-    if (indicators.explicitMarkerCount > 0) codes.push('explicit_marker');
-    return codes;
-  };
-  const displayPriority = (indicators: ExperienceReviewIndicators): ExperienceReviewPriority => {
-    if (indicators.highObservationCount > 0 || indicators.userCorrectionCount > 0 || indicators.userInterruptionCount > 0 || indicators.sessionInterruptedCount > 0 || indicators.negativeFeedbackCount > 0) return 'review_first';
-    if (indicators.mediumObservationCount > 0 || indicators.toolFailureCount > 0 || indicators.hedgingCount > 0 || indicators.explicitMarkerCount > 0 || indicators.hardRuleTextHitCount > 0) return 'sample_review';
-    return 'routine_sample';
-  };
-  const displayInferenceBasisCodes = (indicators: ExperienceReviewIndicators): ExperienceRuleFindingCode[] => {
-    const codes: ExperienceRuleFindingCode[] = [];
-    if (indicators.highObservationCount > 0) codes.push('high_observation_seen');
-    if (indicators.userCorrectionCount > 0) codes.push('user_correction_seen');
-    if (indicators.userInterruptionCount > 0) codes.push('user_interruption_seen');
-    if (indicators.sessionInterruptedCount > 0) codes.push('session_interrupted_seen');
-    if (indicators.negativeFeedbackCount > 0) codes.push('negative_feedback_seen');
-    if (indicators.toolFailureCount > 0) codes.push('tool_failure_seen');
-    if (indicators.mediumObservationCount > 0) codes.push('medium_observation_seen');
-    if (indicators.hedgingCount > 0) codes.push('hedging_seen');
-    if (indicators.explicitMarkerCount > 0) codes.push('explicit_marker_seen');
-    if (indicators.hardRuleTextHitCount > 0) codes.push('hard_rule_seen');
-    if (indicators.positiveFeedbackCount > 0) codes.push('positive_feedback_seen');
-    if (indicators.userGoalShiftCount > 0) codes.push('user_goal_shift_seen');
-    return codes;
-  };
-  const displayAssistiveInference = (
-    indicators: ExperienceReviewIndicators,
-    fallback?: ExperienceAssistiveInference,
-  ): ExperienceAssistiveInference => {
-    const basisRuleCodes = displayInferenceBasisCodes(indicators);
-    const code: ExperienceAssistiveInferenceCode =
-      indicators.highObservationCount > 0 || indicators.userCorrectionCount > 0 || indicators.userInterruptionCount > 0 || indicators.sessionInterruptedCount > 0 || indicators.negativeFeedbackCount > 0
-        ? 'review_recommended'
-        : indicators.mediumObservationCount > 0 || indicators.toolFailureCount > 0 || indicators.hedgingCount > 0 || indicators.explicitMarkerCount > 0 || indicators.hardRuleTextHitCount > 0
-          ? 'sample_recommended'
-          : indicators.positiveFeedbackCount > 0
-            ? 'positive_signal_observed'
-            : indicators.userGoalShiftCount > 0
-              ? 'user_switched_topic_neutral'
-              : 'no_obvious_issue_from_rules';
-    return {
-      mode: 'deterministic_rules_only',
-      code,
-      confidence: fallback?.confidence ?? 'medium',
-      basisRuleCodes,
-      cautionCodes: fallback?.cautionCodes ?? ['no_llm_judge', 'rule_only'],
-      evidenceRefs: fallback?.evidenceRefs ?? [],
-    };
-  };
-  const displayRuleFindings = (
-    session: ExperienceSessionSummary,
-    indicators: ExperienceReviewIndicators,
-  ): ExperienceRuleFinding[] => {
-    const old = session.ruleFindings ?? [];
-    const oldByCode = new Map(old.map((finding) => [finding.code, finding]));
-    const next: ExperienceRuleFinding[] = [];
-    const refsForCode = (code: ExperienceRuleFindingCode): ExperienceEvidenceRef[] | undefined => {
-      const signalType =
-        code === 'user_correction_seen' ? 'correction'
-        : code === 'user_interruption_seen' ? 'interruption'
-        : code === 'negative_feedback_seen' ? 'frustration'
-        : code === 'positive_feedback_seen' ? 'positive'
-        : undefined;
-      if (!signalType) return undefined;
-      const refs = canonicalFeedbackSignalsForDisplay(session, signalType).map((signal) => signal.evidenceRef);
-      return refs.length > 0 ? refs : undefined;
-    };
-    const push = (code: ExperienceRuleFindingCode, level: ExperienceRuleFindingLevel, count: number): void => {
-      if (count <= 0) return;
-      const previous = oldByCode.get(code);
-      next.push({ code, level, count, evidenceRefs: refsForCode(code) ?? previous?.evidenceRefs ?? [] });
-    };
-    push('high_observation_seen', 'attention', indicators.highObservationCount);
-    push('user_correction_seen', 'attention', indicators.userCorrectionCount);
-    push('user_interruption_seen', 'attention', indicators.userInterruptionCount);
-    push('session_interrupted_seen', 'attention', indicators.sessionInterruptedCount);
-    push('negative_feedback_seen', 'attention', indicators.negativeFeedbackCount);
-    push('tool_failure_seen', 'sample', indicators.toolFailureCount);
-    push('medium_observation_seen', 'sample', indicators.mediumObservationCount);
-    push('hedging_seen', 'sample', indicators.hedgingCount);
-    push('explicit_marker_seen', 'sample', indicators.explicitMarkerCount);
-    push('hard_rule_seen', 'sample', indicators.hardRuleTextHitCount);
-    push('positive_feedback_seen', 'normal', indicators.positiveFeedbackCount);
-    push('user_goal_shift_seen', 'normal', indicators.userGoalShiftCount);
-    for (const finding of old) {
-      if ([
-        'high_observation_seen',
-        'user_correction_seen',
-        'user_interruption_seen',
-        'session_interrupted_seen',
-        'negative_feedback_seen',
-        'tool_failure_seen',
-        'medium_observation_seen',
-        'hedging_seen',
-        'explicit_marker_seen',
-        'hard_rule_seen',
-        'positive_feedback_seen',
-        'user_goal_shift_seen',
-        'no_priority_signal',
-      ].includes(finding.code)) continue;
-      next.push(finding);
-    }
-    if (next.filter((finding) => finding.level !== 'normal').length === 0) {
-      next.push({ code: 'no_priority_signal', level: 'normal', count: 1, evidenceRefs: [] });
-    }
-    return next;
-  };
-  const renderSkillEvidenceSummary = (skill: NonNullable<typeof experience>['skills'][number]): string => {
-    const indicators = displayIndicatorsBySkill.get(skill.skillName) ?? skill.indicators;
-    const toolCounts = Object.keys(skill.toolCounts ?? {}).length > 0
-      ? skill.toolCounts
-      : experienceToolCountsBySkill.get(skill.skillName);
-    const displayToolCounts = mergeCountsByLabel(toolCounts, userFacingToolLabel);
-    const total = indicators.toolCallCount;
-    const failed = Math.max(0, indicators.toolFailureCount);
-    const cancelled = Math.max(0, indicators.toolCancelledCount ?? 0);
-    const unknown = Math.max(0, indicators.toolUnknownCount ?? 0);
-    const interrupted = Math.max(0, indicators.userInterruptionCount);
-    const success = Math.max(0, total - failed - cancelled - unknown);
-    const processMetrics = [
-      renderDecisionMetricIfPositive('高优先级', indicators.highObservationCount, 'priority'),
-      renderDecisionMetricIfPositive('抽样过程发现', indicators.mediumObservationCount, 'sample'),
-      renderDecisionMetricIfPositive('用户硬性要求', indicators.hardRuleTextHitCount, 'sample'),
-      renderDecisionMetricIfPositive('自我纠正', indicators.selfCorrectionCount ?? 0, 'sample'),
-      renderDecisionMetricIfPositive('重复执行', indicators.repeatedExecutionCount ?? 0, 'sample'),
-      renderDecisionMetricIfPositive('不确定表达', indicators.hedgingCount, 'sample'),
-      renderDecisionMetricIfPositive('显式缺口', indicators.explicitMarkerCount, 'sample'),
-    ].filter(Boolean).join('');
-    return `<div class="skill-evidence-summary">
-      <div class="summary-row"><span class="summary-title">【skill 运行】</span>${renderMetric('调用段', skill.invocationCount, '次')}<span class="summary-muted">分布 ${skill.sessionCount} 个 session</span>${renderMetric('工具调用', total, '次')}${renderMetricShare('成功', success, total)}${renderDecisionMetricShare('失败', failed, total, 'sample')}${cancelled > 0 ? renderMetricShare('取消', cancelled, total) : ''}${unknown > 0 ? renderMetricShare('状态未知', unknown, total) : ''}${renderDecisionMetric('人工中断', interrupted, 'priority')}${renderDecisionMetricIfPositive('自我纠正', indicators.selfCorrectionCount ?? 0, 'sample')}${renderDecisionMetricIfPositive('重复执行', indicators.repeatedExecutionCount ?? 0, 'sample')}${renderSoftMetric('有结果', indicators.assistantDeliverySignalCount ?? 0, '助手回复里出现明确完成态或结果反馈。')}${renderSoftMetric('有产物', indicators.deliverableArtifactSignalCount ?? 0, '助手回复里出现文档链接、Demo 地址、文件路径、代码块或上传产物。')}</div>
-      <div class="summary-row"><span class="summary-title">【用户交互】</span>${renderDecisionMetric('用户纠正', indicators.userCorrectionCount, 'priority')}${renderDecisionMetric('人工中断', indicators.userInterruptionCount, 'priority')}${renderMetric('追问', indicators.userFollowUpCount)}${renderDecisionMetric('负向反馈', indicators.negativeFeedbackCount ?? 0, 'priority')}${renderMetric('正向反馈', indicators.positiveFeedbackCount ?? 0)}${renderMetric('目标切换', indicators.userGoalShiftCount ?? 0)}</div>
-      <div class="summary-row"><span class="summary-title">【工具调用】</span>${renderMetric('总计', total, '次')}<span class="summary-detail">(${renderRankedCounts(displayToolCounts, total)})</span>${renderDecisionMetricShare('工具执行失败', failed, total, 'sample')}${cancelled > 0 ? renderMetricShare('工具调用取消', cancelled, total) : ''}${unknown > 0 ? renderMetricShare('结果状态未知', unknown, total) : ''}</div>
-      <div class="summary-row"><span class="summary-title">【过程发现】</span>${processMetrics || '<span class="summary-muted">未发现需要展示的过程信号</span>'}</div>
-      ${renderSkillProblemPatterns(skill.skillName, skill.problemPatterns)}
-    </div>`;
-  };
-  const renderSkillChainSummary = (skillName: string): string => {
-    const chain = skillChains[skillName];
-    if (!chain?.definition.found) {
-      return `<span class="summary-muted">未找到 SKILL.md：当前日志能识别调用，但本机目录里没有对应 skill 定义，无法做 doctor 检查。</span>`;
-    }
-    const hard = chain.healthCheck.hardRules;
-    const workflows = chain.healthCheck.workflows;
-    const runtime = chain.runtime.summary;
-    const record = skillDerivedStandards[skillName];
-    const detectedHardCount = (record?.standards ?? []).filter((standard) =>
-      standard.standardKind === 'hard_rule_candidate' && (standard.status === 'author_confirmed' || standard.status === 'pending_review')
-    ).length;
-    const hardText = hard.declared
-      ? `标准化规则 ${hard.count} 条`
-      : detectedHardCount > 0
-        ? `规则检测 ${detectedHardCount} 条`
-        : '标准化规则未声明';
-    const workflowText = workflows.workflows.length > 0
-      ? workflows.declared
-        ? `标准化流程 ${workflows.branchCount} 条 / 节点 ${workflows.nodeCount}`
-        : workflows.source === 'markdown_headings'
-          ? `流程检测 ${workflows.branchCount} 条 / 节点 ${workflows.nodeCount}（正文流程）`
-          : `流程检测 ${workflows.branchCount} 条 / 节点 ${workflows.nodeCount}`
-      : '标准化流程未声明';
-    const pending = (record?.standards ?? []).filter((standard) => standard.status === 'pending_review' || standard.status === 'stale');
-    const hardPending = pending.filter((standard) => standard.standardKind === 'hard_rule_candidate');
-    const workflowPending = pending.filter((standard) => standard.standardKind === 'workflow_candidate');
-    const pendingLine = pending.length > 0
-      ? `<div class="skill-chain-compact-candidates">
-          <span>待确认 ${pending.length} 条</span>
-          ${hardPending[0] ? `<span title="${e(hardPending[0].title)}">规则：${e(truncateText(hardPending[0].title, 18))}</span>` : ''}
-          ${workflowPending[0] ? `<span title="${e(workflowPending[0].title)}">流程：${e(truncateText(workflowPending[0].title, 18))}</span>` : ''}
-        </div>`
-      : '';
-    return `<div class="skill-chain-cell-summary">
-      <div>${e(hardText)} · ${e(workflowText)}</div>
-      <div>证据符合 ${runtime.passedCount} / 需关注 ${runtime.attentionCount} / 无法判断 ${runtime.manualReviewCount}</div>
-      ${pendingLine}
-    </div>`;
-  };
-  const collectSkillChainAdvisoryCodes = (skillName: string): SkillChainAdvisoryCode[] => {
-    const chain = skillChains[skillName];
-    if (!chain) return [];
-    if (!chain.definition.found) return ['skill_md_not_found'];
-    const codes: SkillChainAdvisoryCode[] = [];
-    if (chain.healthCheck.hardRules.advisoryCode) codes.push(chain.healthCheck.hardRules.advisoryCode);
-    if (chain.healthCheck.workflows.advisoryCode) codes.push(chain.healthCheck.workflows.advisoryCode);
-    return codes;
-  };
-  const renderSkillChainButton = (skillName: string, templateId: string): string => {
-    const chain = skillChains[skillName];
-    const advisoryCodes = collectSkillChainAdvisoryCodes(skillName);
-    const advisoryCount = advisoryCodes.length;
-    const hasAdvisory = advisoryCount > 0;
-    const advisoryLabels = advisoryCodes.map((code) => getSkillChainAdvisory(code).shortLabel);
-    const title = chain?.definition.found
-      ? (hasAdvisory
-        ? `点开查看 skill 定义、标准规则和运行时证据；当前缺：${advisoryLabels.join('、')}。`
-        : '点开查看 skill 定义、标准规则和运行时证据。')
-      : '本机目录里没有这个 skill 的 SKILL.md。点开看建议怎么补，或试试 omk doctor。';
-    return `<button type="button" class="context-chain-button${hasAdvisory ? ' has-advisory' : ''}" onclick="event.stopPropagation(); openContextChainModal('${e(templateId)}', this)" title="${e(title)}"><span class="context-chain-button-icon" aria-hidden="true">🔗</span><span class="context-chain-button-main">定义链路</span>${hasAdvisory ? `<span class="context-chain-button-advisory-list">${advisoryLabels.map((label) => `<span class="context-chain-button-advisory">${e(label)}</span>`).join('')}</span>` : '<span class="context-chain-button-ok">标准已声明</span>'}</button>`;
-  };
-  const runtimeLiteStatusIcon = (status: string): string =>
-    status === 'passed' ? '✅' : status === 'attention' ? '❌' : '?';
-  const runtimeLiteStatusLabel = (status: string): string =>
-    status === 'passed' ? '有证据' : status === 'attention' ? '需复核' : '待判断';
-  const displayRuntimeLiteTitle = (title: string, id: string, index: number): string => {
-    const source = title || id;
-    const parts = source.split('/').map((part) => part.trim()).filter(Boolean);
-    const stepLabel = (value: string): string => {
-      const match = value.match(/step[\s_-]*([0-9]+)/i);
-      return match ? `第 ${match[1]} 步` : `第 ${index + 1} 步`;
-    };
-    if (parts.length >= 2 && parts[0] === 'markdown_steps') return `正文流程节点 / ${stepLabel(parts[1])}`;
-    if (/^markdown_steps[._-]step/i.test(id)) return `正文流程节点 / ${stepLabel(id.replace(/^markdown_steps[._-]/i, ''))}`;
-    if (/^step[\s_-]*[0-9]+$/i.test(source)) return stepLabel(source);
-    return source || `检查项 ${index + 1}`;
-  };
-  const shouldShowRuntimeLiteRawId = (id: string): boolean =>
-    Boolean(id) && !/^markdown_steps[._-]step/i.test(id) && id !== 'markdown_steps';
-  const runtimeStepDepth = (title: string, id: string): number => {
-    const source = `${id} ${title}`.toLowerCase();
-    if (/\bstep[\s_-]*\d+[a-z]\b/.test(source)) return 1;
-    if (/\bstep[\s_-]*\d+[._-][a-z0-9]+\b/.test(source)) return 1;
-    return 0;
-  };
-  const renderRuntimeLiteList = (
-    checks: Array<{ id: string; title: string; status: string; reason?: string; evidenceSnippets: string[] }>,
-    kind: 'workflow' | 'rule',
-    emptyText = '暂无可检查项。',
-  ): string => checks.length > 0
-    ? `<ol class="runtime-step-list">${checks.map((check, index) => `<li class="runtime-step runtime-${e(check.status)} is-depth-${runtimeStepDepth(check.title, check.id)}">
-        <div class="runtime-step-head">
-          <span class="runtime-step-index">${kind === 'workflow' ? `第 ${index + 1} 步` : `规则 ${index + 1}`}</span>
-          <span class="runtime-step-name">${e(displayRuntimeLiteTitle(check.title, check.id, index))}</span>
-          <span class="runtime-step-state" title="${e(runtimeLiteStatusLabel(check.status))}">${runtimeLiteStatusIcon(check.status)}</span>
-        </div>
-        <details class="runtime-step-detail">
-          <summary>查看证据</summary>
-          <div class="runtime-step-detail-body">
-            ${shouldShowRuntimeLiteRawId(check.id) ? `<code>原始编号：${e(check.id)}</code>` : ''}
-            ${check.reason ? `<p>${e(check.reason)}</p>` : ''}
-            ${check.evidenceSnippets.length > 0 ? `<ul>${check.evidenceSnippets.map((snippet) => `<li>${e(snippet)}</li>`).join('')}</ul>` : '<p>没有可展示的证据片段。</p>'}
-          </div>
-        </details>
-      </li>`).join('')}</ol>`
-    : emptyText ? `<p class="context-muted">${e(emptyText)}</p>` : '';
-  const renderRuntimeRuleFlow = (skillName: string): string => {
-    const chain = skillChains[skillName];
-    const enhancedReview = skillDerivedStandards[skillName]?.enhancedReview;
-    const extractedStandards = enhancedReview?.extractedStandards;
-    if (!chain && !extractedStandards) return '<p class="context-muted">没有找到这个 skill 的流程规则检测结果。</p>';
-    const workflows = chain?.healthCheck.workflows.workflows ?? [];
-    const hardRules = chain?.healthCheck.hardRules.rules ?? [];
-    const llmWorkflowStandards = [
-      ...(extractedStandards?.workflows ?? []),
-      ...(extractedStandards?.completionCriteria ?? []),
-      ...(extractedStandards?.artifactCriteria ?? []),
-    ];
-    const llmHardRuleStandards = extractedStandards?.hardrules ?? [];
-    const llmStandardNodes = extractedStandards?.standardNodes ?? [];
-    const llmWorkflowNodes = llmStandardNodes.filter((node) => node.nodeKind !== 'hardRule');
-    const llmHardRuleNodes = llmStandardNodes.filter((node) => node.nodeKind === 'hardRule');
-    const standardNodeById = new Map(llmStandardNodes.map((node) => [node.nodeId, node] as const));
-    const standardNodeParentById = new Map<string, string>();
-    for (const node of llmStandardNodes) {
-      for (const childId of node.childNodeIds ?? []) {
-        if (standardNodeById.has(childId) && !standardNodeParentById.has(childId)) {
-          standardNodeParentById.set(childId, node.nodeId);
-        }
-      }
-    }
-    const standardNodeDepth = (node: typeof llmStandardNodes[number]): number => {
-      let depth = 0;
-      let current = standardNodeParentById.get(node.nodeId);
-      const visited = new Set<string>([node.nodeId]);
-      while (current && !visited.has(current) && depth < 3) {
-        depth += 1;
-        visited.add(current);
-        current = standardNodeParentById.get(current);
-      }
-      if (depth > 0) return depth;
-      return runtimeStepDepth(node.title, node.nodeId);
-    };
-    const sourceContent = chain?.definition.content ?? '';
-    const sourcePreview = sourceContent.length > 12000 ? `${sourceContent.slice(0, 12000)}\n\n... 已截断，仅展示前 12000 字符` : sourceContent;
-    const runtimeNodeResultById = new Map((enhancedReview?.runtimeNodeResults?.nodes ?? [])
-      .map((node) => [node.nodeId, node] as const));
-    const modelNodeStatusText = (status?: string): string => {
-      if (status === 'passed') return '日志命中';
-      if (status === 'failed' || status === 'missed') return '未发现调用';
-      if (status === 'violated') return '命中异常';
-      if (status === 'degraded') return '证据不可信';
-      if (status === 'not_applicable') return '不适用';
-      if (status === 'unknown') return '未发现调用';
-      return '未发现调用';
-    };
-    const modelNodeStatusClass = (status?: string): string => {
-      if (status === 'passed') return 'is-hit';
-      if (status === 'failed' || status === 'missed' || status === 'violated' || status === 'degraded') return 'is-attention';
-      return '';
-    };
-    const modelNodeStatusIcon = (status?: string): string => {
-      if (status === 'passed') return '✅';
-      if (status === 'failed' || status === 'violated' || status === 'degraded') return '❌';
-      return '?';
-    };
-    type RuntimeNodeReview = NonNullable<SkillLlmEnhancedReviewSections['runtimeNodeResults']>['nodes'][number]
-      | NonNullable<SkillLlmEnhancedReviewSections['runtimeNodeAssessment']>['nodes'][number];
-    const reviewEvidenceSnippets = (review: RuntimeNodeReview): string[] => {
-      if ('evidenceRefs' in review) return (review.evidenceRefs ?? []).map((ref) => ref.snippet).filter((snippet): snippet is string => Boolean(snippet)).slice(0, 3);
-      return (review.evidence ?? []).slice(0, 3);
-    };
-    const compactRuntimeEvidence = (body: string, summary = '展开关键证据'): string => `<details class="runtime-rule-node-model-detail">
-      <summary>${e(summary)}</summary>
-      <div>${body}</div>
-    </details>`;
-    const renderRuntimeNodeReview = (review?: RuntimeNodeReview, fallback?: { reason?: string; evidenceSnippets: string[] }): string => {
-      if (!review && !fallback) return `<div class="runtime-rule-node-model is-muted">
-        <strong>未发现调用</strong>
-        ${compactRuntimeEvidence('<p>运行日志中没有命中这个节点要求的调用信号。</p>', '展开说明')}
-      </div>`;
-      if (!review) return `<div class="runtime-rule-node-model">
-        <strong>规则命中</strong>
-        ${compactRuntimeEvidence(`${fallback?.reason ? `<p>${e(fallback.reason)}</p>` : ''}${(fallback?.evidenceSnippets ?? []).length > 0 ? `<ul>${(fallback?.evidenceSnippets ?? []).slice(0, 3).map((snippet) => `<li>${e(snippet)}</li>`).join('')}</ul>` : '<p>没有可展示的证据片段。</p>'}`)}
-      </div>`;
-      const snippets = reviewEvidenceSnippets(review);
-      return `<div class="runtime-rule-node-model runtime-node-model-${e(review.status)}">
-        <strong>${e(modelNodeStatusText(review.status))}</strong>
-        ${compactRuntimeEvidence(`${review.reason ? `<p>${e(review.reason)}</p>` : ''}${snippets.length > 0 ? `<ul>${snippets.map((snippet) => `<li>${e(snippet)}</li>`).join('')}</ul>` : '<p>没有可展示的证据片段。</p>'}`)}
-      </div>`;
-    };
-    const renderSourceHints = (node: typeof llmStandardNodes[number]): string => {
-      const hints = (node.sourceHints ?? []).slice(0, 3);
-      if (hints.length === 0) return '<p>来源原文：模型抽取节点，未返回可定位原文。</p>';
-      return `<div class="runtime-rule-source-hints">
-        <span>来源原文</span>
-        <ul>${hints.map((hint) => `<li>${hint.line ? `<em>第 ${e(String(hint.line))} 行</em>` : ''}${e(hint.snippet)}</li>`).join('')}</ul>
-      </div>`;
-    };
-    const signalLabel = (signal: typeof llmStandardNodes[number]['expectedSignals'][number]): string => {
-      const typeLabel: Record<string, string> = {
-        tool_name: '工具',
-        tool_input: '工具输入',
-        tool_output: '工具输出',
-        assistant_text: '助手文本',
-        user_text: '用户文本',
-        artifact_kind: '产物类型',
-        artifact_path: '产物路径',
-        event_kind: '事件',
-        file_path: '文件路径',
-      };
-      return `${typeLabel[signal.type] ?? signal.type}:${Array.isArray(signal.value) ? signal.value.join('/') : signal.value}`;
-    };
-    const renderSignalSummary = (node: typeof llmStandardNodes[number]): string => {
-      const expected = node.expectedSignals.slice(0, 3).map(signalLabel);
-      const forbidden = node.forbiddenSignals.slice(0, 2).map(signalLabel);
-      const failure = node.failureSignals.slice(0, 2).map(signalLabel);
-      const parts = [
-        expected.length > 0 ? `期望 ${expected.join('、')}` : '',
-        forbidden.length > 0 ? `禁止 ${forbidden.join('、')}` : '',
-        failure.length > 0 ? `失败信号 ${failure.join('、')}` : '',
-      ].filter(Boolean);
-      return parts.length > 0
-        ? `<p class="runtime-rule-match-spec">日志匹配口径：${e(parts.join('；'))}</p>`
-        : '<p class="runtime-rule-match-spec">日志匹配口径：未抽取到结构化信号。</p>';
-    };
-    const renderModelStandards = (items: Array<{ title: string; body: string; evidence?: string[] }>): string =>
-      items.length > 0
-        ? `<div class="runtime-rule-model-list">
-          <em>模型识别的候选标准，尚未等同于运行命中</em>
-          ${items.map((item) => `<article class="runtime-rule-breakdown-item is-model">
-            <strong>${e(item.title)}</strong>
-            ${item.body ? `<p>${e(item.body)}</p>` : ''}
-            ${(item.evidence ?? []).length > 0 ? `<ul>${(item.evidence ?? []).slice(0, 3).map((snippet) => `<li>${e(snippet)}</li>`).join('')}</ul>` : ''}
-          </article>`).join('')}
-        </div>`
-        : '';
-    const renderStandardNodeList = (nodes: typeof llmStandardNodes): string =>
-      nodes.length > 0
-        ? `<ol class="runtime-rule-node-list">${nodes.map((node, index) => {
-          const depth = standardNodeDepth(node);
-          return `<li class="runtime-rule-node is-depth-${depth}">
-            <div class="runtime-rule-node-head">
-              <span>${node.nodeKind === 'stage' ? `阶段 ${index + 1}` : `节点 ${index + 1}`}</span>
-              <strong>${e(node.title)}</strong>
-              <em>标准拆解</em>
-            </div>
-            ${node.description ? `<p>${e(node.description)}</p>` : ''}
-            ${renderSourceHints(node)}
-            ${renderSignalSummary(node)}
-          </li>`;
-        }).join('')}</ol>`
-        : '';
-    const renderRuntimeNodeResultList = (nodes: typeof llmStandardNodes): string =>
-      nodes.length > 0
-        ? `<ol class="runtime-step-list runtime-rule-execution-list">${nodes.map((node, index) => {
-          const review = runtimeNodeResultById.get(node.nodeId);
-          const depth = standardNodeDepth(node);
-          return `<li class="runtime-step runtime-rule-execution-item ${modelNodeStatusClass(review?.status)} is-depth-${depth}">
-            <div class="runtime-step-head">
-              <span class="runtime-step-index">${node.nodeKind === 'stage' ? `阶段 ${index + 1}` : `节点 ${index + 1}`}</span>
-              <span class="runtime-step-name">${e(node.title)}</span>
-              <span class="runtime-step-state" title="${e(modelNodeStatusText(review?.status))}">${modelNodeStatusIcon(review?.status)}</span>
-            </div>
-            ${renderRuntimeNodeReview(review)}
-          </li>`;
-        }).join('')}</ol>`
-        : '';
-    const workflowBreakdown = workflows.length > 0
-      ? workflows.map((workflow, workflowIndex) => `<div class="runtime-rule-flow-block">
-          <div class="runtime-rule-flow-title">${e(workflow.id === 'markdown_steps' ? '正文流程' : `流程 ${workflowIndex + 1}`)}${workflow.description ? `<span>${e(workflow.description)}</span>` : ''}</div>
-          <ol class="runtime-rule-node-list">${workflow.nodes.map((node, nodeIndex) => {
-            return `<li class="runtime-rule-node">
-              <div class="runtime-rule-node-head">
-                <span>第 ${nodeIndex + 1} 步</span>
-                <strong>${e(node.action || `流程节点 ${nodeIndex + 1}`)}</strong>
-                <em>原文映射</em>
-              </div>
-              <p>来源原文：SKILL.md ${workflow.id === 'markdown_steps' ? '正文步骤' : `workflow ${workflow.id}`}。</p>
-              <p class="runtime-rule-match-spec">日志匹配口径：根据这一步的动作关键词和工具调用证据进行匹配。</p>
-            </li>`;
-          }).join('')}</ol>
-        </div>`).join('')
-      : `<p class="context-muted">${chain?.healthCheck.workflows.declared ? '未解析到流程节点。' : '未标准化声明流程，当前只能依赖正文探测或运行时推断。'}</p>`;
-    const hardRuleBreakdown = hardRules.length > 0
-      ? `<ol class="runtime-rule-node-list">${hardRules.map((rule, ruleIndex) => {
-        return `<li class="runtime-rule-node">
-          <div class="runtime-rule-node-head">
-            <span>规则 ${ruleIndex + 1}</span>
-            <strong>${e(rule.rule)}</strong>
-            <em>原文映射</em>
-          </div>
-          ${rule.expectedBehavior ? `<p>来源原文：${e(rule.expectedBehavior)}</p>` : '<p>来源原文：SKILL.md hardRules 声明。</p>'}
-          <p class="runtime-rule-match-spec">日志匹配口径：根据规则约束和异常/禁止信号进行匹配。</p>
-        </li>`;
-      }).join('')}</ol>`
-      : `<p class="context-muted">${chain?.healthCheck.hardRules.declared ? '未解析到硬性规则。' : '未标准化声明硬性规则，当前只能展示已探测到的规则证据。'}</p>`;
-    const renderWorkflowRuntimeList = (): string => {
-      const deterministic = chain ? renderRuntimeLiteList(chain.runtime.workflowNodes, 'workflow', '') : '';
-      const model = renderRuntimeNodeResultList(llmWorkflowNodes);
-      const content = `${deterministic}${model}`;
-      return content || '<p class="context-muted">没有规则层流程命中数据。</p>';
-    };
-    const renderHardRuleRuntimeList = (): string => {
-      const deterministic = chain ? renderRuntimeLiteList(chain.runtime.hardRules, 'rule', '') : '';
-      const model = renderRuntimeNodeResultList(llmHardRuleNodes);
-      const content = `${deterministic}${model}`;
-      return content || '<p class="context-muted">没有规则层规则命中数据。</p>';
-    };
-    return `<div class="runtime-rule-three-col">
-      <section class="runtime-rule-column">
-        <h5>Skill 原文</h5>
-        ${chain?.definition.found
-          ? `<div class="runtime-rule-source-path"><code>${e(chain.definition.path ?? '')}</code></div><pre class="runtime-rule-source">${e(sourcePreview)}</pre>`
-          : '<p class="context-muted">未找到本地 SKILL.md，只能展示运行时证据。</p>'}
-      </section>
-      <section class="runtime-rule-column">
-        <h5>标准拆解 / 原文映射</h5>
-        <p class="runtime-rule-column-hint">这一列只说明抽象节点来自哪段 SKILL.md，以及后续用什么口径去匹配日志。</p>
-        <div class="runtime-rule-column-group">
-          <h6>流程</h6>
-          ${workflowBreakdown}
-          ${renderStandardNodeList(llmWorkflowNodes)}
-          ${renderModelStandards(llmWorkflowStandards)}
-        </div>
-        <div class="runtime-rule-column-group">
-          <h6>硬性规则</h6>
-          ${hardRuleBreakdown}
-          ${renderStandardNodeList(llmHardRuleNodes)}
-          ${renderModelStandards(llmHardRuleStandards)}
-        </div>
-      </section>
-      <section class="runtime-rule-column">
-        <h5>日志命中 / 执行判断</h5>
-        <p class="runtime-rule-column-hint">这一列只看真实运行日志：节点有没有命中、是否异常，以及对应证据。</p>
-        <div class="runtime-rule-column-group">
-          <h6>流程命中</h6>
-          ${renderWorkflowRuntimeList()}
-        </div>
-        <div class="runtime-rule-column-group">
-          <h6>规则命中</h6>
-          ${renderHardRuleRuntimeList()}
-        </div>
-      </section>
-    </div>`;
-  };
-  const renderSkillChainTemplate = (skillName: string): string => {
-    const chain = skillChains[skillName];
-    const renderAdvisoryBlock = (advisoryCode?: SkillChainAdvisoryCode, skillNameForCmd?: string): string => {
-      if (!advisoryCode) return '';
-      const advisory = getSkillChainAdvisory(advisoryCode);
-      const exampleBlock = advisory.exampleYaml
-        ? `<details class="skill-chain-advisory-example">
-            <summary>查看建议示例</summary>
-            <pre>${e(advisory.exampleYaml)}</pre>
-          </details>`
-        : '';
-      const resolvedCommand = skillNameForCmd ? resolveAdvisoryCommand(advisory, skillNameForCmd) : undefined;
-      const commandBlock = resolvedCommand
-        ? `<div class="skill-chain-advisory-cmd-wrap">
-            <div class="skill-chain-advisory-cmd-label">如需更多体检维度（依赖检查、可读性等），可以试试 omk doctor（非必须）：</div>
-            <div class="skill-chain-advisory-cmd-row">
-              <code class="skill-chain-advisory-cmd" data-copy-target="${e(resolvedCommand)}">${e(resolvedCommand)}</code>
-              <button type="button" class="skill-chain-advisory-copy-btn" data-copy-source="${e(resolvedCommand)}" title="复制命令到剪贴板">复制</button>
-            </div>
-          </div>`
-        : '';
-      return `<div class="skill-chain-advisory">
-        <div class="skill-chain-advisory-message">⚠️ ${e(advisory.message)}</div>
-        ${exampleBlock}
-        ${commandBlock}
-      </div>`;
-    };
-    if (!chain) {
-      // SKILL.md 找不到也是一种 advisory，统一用 advisory block 渲染；
-      // 同时给出（非强制的）omk doctor 建议命令。
-      const notFoundAdvisory = renderAdvisoryBlock('skill_md_not_found', skillName);
-      return `<div class="context-chain-grid">
-        <section class="context-chain-panel">
-          <h3>① 原始 SKILL.md</h3>
-          ${notFoundAdvisory}
-        </section>
-        <section class="context-chain-panel">
-          <h3>② 规则 / 流程探测</h3>
-          <p class="context-muted">缺少本地 SKILL.md，当前无法探测规则和流程。这不代表作者没声明，只是本地看不到。</p>
-        </section>
-        <section class="context-chain-panel">
-          <h3>③ 运行时</h3>
-          <p class="context-muted">运行时证据可从 observe 时间线查看；静态定义链路在本地暂不可用。</p>
-        </section>
-        <section class="context-chain-panel">
-          <h3>④ 候选确认</h3>
-          <p class="context-muted">暂无可确认候选。</p>
-        </section>
-      </div>`;
-    }
-    const hard = chain.healthCheck.hardRules;
-    const workflows = chain.healthCheck.workflows;
-    const record = skillDerivedStandards[skillName];
-    const softStandards = record?.standards ?? [];
-    const statusPriority: Record<SkillDerivedStandard['status'], number> = {
-      pending_review: 0,
-      author_confirmed: 1,
-      rejected: 2,
-      stale: 3,
-    };
-    const sortedSoftStandards = [...softStandards].sort((a, b) =>
-      statusPriority[a.status] - statusPriority[b.status] || a.standardKind.localeCompare(b.standardKind) || a.title.localeCompare(b.title)
-    );
-    const kindLabel = (kind: SkillDerivedStandard['standardKind']): string =>
-      kind === 'workflow_candidate' ? '流程候选' : '规则候选';
-    const annotationStateClass = (status: SkillDerivedStandard['status']): string =>
-      status === 'author_confirmed' ? 'is-confirmed' : status === 'rejected' ? 'is-rejected' : '';
-    const annotationStateIcon = (status: SkillDerivedStandard['status']): string =>
-      status === 'author_confirmed' ? '✅' : status === 'rejected' ? '❌' : '';
-    const candidateSourceTexts = (standard: SkillDerivedStandard): string[] => [
-      ...standard.evidence,
-      standard.title,
-      standard.body,
-    ]
-      .map((text) => text.trim())
-      .filter((text) => text.length >= 4)
-      .sort((a, b) => b.length - a.length);
-    const findCandidateRange = (
-      content: string,
-      standard: SkillDerivedStandard,
-      usedRanges: Array<{ start: number; end: number }>,
-    ): { start: number; end: number; text: string } | undefined => {
-      for (const text of candidateSourceTexts(standard)) {
-        const start = content.indexOf(text);
-        const end = start + text.length;
-        if (start >= 0 && !usedRanges.some((range) => start < range.end && end > range.start)) {
-          return { start, end, text };
-        }
-      }
-      return undefined;
-    };
-    const renderCandidateActions = (standard: SkillDerivedStandard): string => `
-      <span class="skill-md-annotation-actions">
-        <span data-soft-standard-status="${e(standard.status)}">${e(softStandardStatusLabel(standard.status))}</span>
-        <button type="button" data-soft-standard-action="author_confirmed" onclick="setSoftStandardStatus('${e(skillName)}', '${e(standard.id)}', 'author_confirmed', this)">确认</button>
-        <button type="button" data-soft-standard-action="rejected" onclick="setSoftStandardStatus('${e(skillName)}', '${e(standard.id)}', 'rejected', this)">否决</button>
-      </span>`;
-    const renderAnnotatedSkillMd = (content: string): string => {
-      if (sortedSoftStandards.length === 0) return e(content);
-      const usedRanges: Array<{ start: number; end: number }> = [];
-      const annotations: Array<{ standard: SkillDerivedStandard; range: { start: number; end: number; text: string } }> = [];
-      for (const standard of sortedSoftStandards) {
-        const range = findCandidateRange(content, standard, usedRanges);
-        if (!range) continue;
-        usedRanges.push(range);
-        annotations.push({ standard, range });
-      }
-      annotations.sort((a, b) => a.range.start - b.range.start);
-      if (annotations.length === 0) return e(content);
-      let cursor = 0;
-      const parts: string[] = [];
-      for (const { standard, range } of annotations) {
-        parts.push(e(content.slice(cursor, range.start)));
-        parts.push(`<span class="skill-md-highlight skill-md-highlight-${standard.standardKind === 'workflow_candidate' ? 'workflow' : 'rule'}" data-soft-standard-id="${e(standard.id)}">${e(content.slice(range.start, range.end))}</span><span class="skill-md-annotation ${annotationStateClass(standard.status)}" data-soft-standard-id="${e(standard.id)}" data-soft-standard-skill="${e(skillName)}"><span class="skill-md-annotation-icon" data-soft-standard-icon="${e(standard.status)}">${e(annotationStateIcon(standard.status))}</span><span class="skill-md-annotation-content"><strong>${e(kindLabel(standard.standardKind))}：</strong>${e(standard.title)}${renderCandidateActions(standard)}</span></span>`);
-        cursor = range.end;
-      }
-      parts.push(e(content.slice(cursor)));
-      return parts.join('');
-    };
-    const unlocatedStandards = sortedSoftStandards.filter((standard) => {
-      const content = chain.definition.content ?? '';
-      return !candidateSourceTexts(standard).some((text) => content.includes(text));
-    });
-    const renderUnlocatedCandidates = (): string => unlocatedStandards.length > 0
-      ? `<details class="skill-md-unlocated">
-          <summary>未定位到原文的候选 ${unlocatedStandards.length} 条</summary>
-          <div class="soft-standard-modal-list">${unlocatedStandards.map((standard) => `<div class="soft-standard-modal-item ${annotationStateClass(standard.status)}" data-soft-standard-id="${e(standard.id)}" data-soft-standard-skill="${e(skillName)}">
-            <div class="soft-standard-modal-head">
-              <span class="skill-md-annotation-icon" data-soft-standard-icon="${e(standard.status)}">${e(annotationStateIcon(standard.status))}</span>
-              <strong>${e(kindLabel(standard.standardKind))}：${e(standard.title)}</strong>
-              <span data-soft-standard-status="${e(standard.status)}">${e(softStandardStatusLabel(standard.status))}</span>
-            </div>
-            <div class="soft-standard-modal-body">${e(standard.body)}</div>
-            ${standard.evidence.length > 0 ? `<div class="soft-standard-modal-evidence">依据：${standard.evidence.map((entry) => e(entry)).join('；')}</div>` : ''}
-            <div class="soft-standard-actions">
-              <button type="button" data-soft-standard-action="author_confirmed" onclick="setSoftStandardStatus('${e(skillName)}', '${e(standard.id)}', 'author_confirmed', this)">确认</button>
-              <button type="button" data-soft-standard-action="rejected" onclick="setSoftStandardStatus('${e(skillName)}', '${e(standard.id)}', 'rejected', this)">否决</button>
-            </div>
-          </div>`).join('')}</div>
-        </details>`
-      : '';
-    const missingNotices = [
-      !hard.declared ? renderAdvisoryBlock(hard.advisoryCode, skillName) : '',
-      !workflows.declared ? renderAdvisoryBlock(workflows.advisoryCode, skillName) : '',
-    ].filter(Boolean).join('');
-    const detectedHardRuleStandards = sortedSoftStandards.filter((standard) =>
-      standard.standardKind === 'hard_rule_candidate' && (standard.status === 'author_confirmed' || standard.status === 'pending_review')
-    );
-    const detectedWorkflowStandards = sortedSoftStandards.filter((standard) =>
-      standard.standardKind === 'workflow_candidate' && (standard.status === 'author_confirmed' || standard.status === 'pending_review')
-    );
-    const detectedMarker = (status: SkillDerivedStandard['status']): string =>
-      status === 'author_confirmed' ? '✅' : '待';
-    const renderStructuredHardRules = (): string => hard.rules.length > 0
-      ? `<div class="standard-checklist">${hard.rules.map((rule) => `<article class="standard-check-item">
-          <div class="standard-check-marker" aria-hidden="true">✓</div>
-          <div class="standard-check-body">
-            <div class="standard-check-title"><code>${e(rule.id)}</code><strong>${e(rule.rule)}</strong></div>
-            <div class="standard-check-expectation">期望行为：${e(rule.expectedBehavior)}</div>
-          </div>
-        </article>`).join('')}</div>`
-      : `<div class="probe-anomaly">
-          <strong>未声明标准化 hardRules</strong>
-          <span>以下内容为规则检测结果；建议优化为能力定义文件顶部声明的标准化规则。</span>
-        </div>
-        ${detectedHardRuleStandards.length > 0
-          ? `<div class="standard-checklist">${detectedHardRuleStandards.map((standard) => `<article class="standard-check-item">
-              <div class="standard-check-marker" aria-hidden="true">${e(detectedMarker(standard.status))}</div>
-              <div class="standard-check-body">
-                <div class="standard-check-title"><code>${e(standard.id)}</code><strong>${e(standard.title)}</strong></div>
-                <div class="standard-check-expectation">${e(standard.body)}</div>
-              </div>
-            </article>`).join('')}</div>`
-          : '<p class="context-muted">暂未检测到可展示的规则结果。</p>'}`;
-    const workflowStandardNotice = !workflows.declared
-      ? `<div class="probe-anomaly">
-          <strong>未声明标准化 workflows</strong>
-          <span>以下内容为流程检测结果；建议优化为能力定义文件顶部声明的标准化流程。</span>
-        </div>`
-      : '';
-    const displayStepLabel = (id: string, index: number): string => {
-      const match = id.match(/^step[\s_-]*([0-9]+)$/i);
-      if (match) return `第 ${match[1]} 步`;
-      return id ? `流程节点 ${index + 1}` : `第 ${index + 1} 步`;
-    };
-    const displayWorkflowLabel = (id: string, index: number): string =>
-      id === 'markdown_steps' ? '正文流程节点' : `流程 ${index + 1}`;
-    const displayRuntimeCheckTitle = (title: string, id: string, index: number): string => {
-      const source = title || id;
-      const parts = source.split('/').map((part) => part.trim()).filter(Boolean);
-      if (parts.length >= 2 && parts[0] === 'markdown_steps') return `正文流程节点 / ${displayStepLabel(parts[1], index)}`;
-      if (/^markdown_steps[._-]step/i.test(id)) return `正文流程节点 / ${displayStepLabel(id.replace(/^markdown_steps[._-]/i, ''), index)}`;
-      if (/^step[\s_-]*[0-9]+$/i.test(source)) return displayStepLabel(source, index);
-      return source || `检查项 ${index + 1}`;
-    };
-    const shouldShowRuntimeCheckRawId = (id: string): boolean =>
-      Boolean(id) && !/^markdown_steps[._-]step/i.test(id) && id !== 'markdown_steps';
-    const renderStructuredWorkflows = (): string => workflows.workflows.length > 0
-      ? `${workflowStandardNotice}<div class="workflow-line-list">${workflows.workflows.map((workflow, workflowIndex) => `<article class="workflow-line">
-          <div class="workflow-line-head">
-            <strong>${e(displayWorkflowLabel(workflow.id, workflowIndex))}</strong>
-            <span>${workflow.source === 'markdown_headings' ? '正文流程探测' : '标准化声明'} · ${workflow.nodes.length} 个节点</span>
-          </div>
-          ${workflow.description ? `<div class="workflow-line-desc">${e(workflow.description)}</div>` : ''}
-          <div class="workflow-node-line">${workflow.nodes.map((node, index) => `<div class="workflow-node">
-            <div class="workflow-node-index">${index + 1}</div>
-            <div class="workflow-node-card">
-              <strong>${e(displayStepLabel(node.id, index))}</strong>
-              <span>${e(node.action)}</span>
-            </div>
-          </div>`).join('')}</div>
-        </article>`).join('')}</div>`
-      : `<div class="probe-anomaly">
-          <strong>未声明标准化 workflows</strong>
-          <span>以下内容为流程检测结果；建议优化为能力定义文件顶部声明的标准化流程。</span>
-        </div>
-        ${detectedWorkflowStandards.length > 0
-          ? `<div class="standard-checklist">${detectedWorkflowStandards.map((standard) => `<article class="standard-check-item">
-              <div class="standard-check-marker" aria-hidden="true">${e(detectedMarker(standard.status))}</div>
-              <div class="standard-check-body">
-                <div class="standard-check-title"><code>${e(standard.id)}</code><strong>${e(standard.title)}</strong></div>
-                <div class="standard-check-expectation">${e(standard.body)}</div>
-              </div>
-            </article>`).join('')}</div>`
-          : '<p class="context-muted">暂未检测到可展示的流程结果。</p>'}`;
-    const renderProbeLog = (): string => {
-      const hardErrors = hard.errors.length > 0 ? hard.errors.join('；') : '无';
-      const workflowErrors = workflows.errors.length > 0 ? workflows.errors.join('；') : '无';
-      const confirmedHardRules = detectedHardRuleStandards.filter((standard) => standard.status === 'author_confirmed').length;
-      const hardDeclareText = hard.declared
-        ? `已标准化声明 ${hard.count} 条`
-        : detectedHardRuleStandards.length > 0
-          ? `未标准化声明；已有规则检测结果 ${detectedHardRuleStandards.length} 条，其中人工确认 ${confirmedHardRules} 条`
-          : '未标准化声明；暂无规则检测结果';
-      const workflowDeclareText = workflows.source === 'frontmatter'
-        ? `已标准化声明 ${workflows.branchCount} 条链路 / ${workflows.nodeCount} 个节点`
-        : workflows.source === 'markdown_headings'
-          ? `未标准化声明；已从正文流程标题探测 ${workflows.branchCount} 条链路 / ${workflows.nodeCount} 个节点`
-          : '未声明';
-      return `<details class="probe-log" ${hard.rules.length === 0 || workflows.workflows.length === 0 ? 'open' : ''}>
-        <summary>原始探测信息</summary>
-        <table>
-          <tbody>
-            <tr><th>探测来源</th><td>${e(chain.healthCheck.source)}</td></tr>
-            <tr><th>规则声明</th><td>${e(hardDeclareText)}</td></tr>
-            <tr><th>规则异常</th><td>${e(hardErrors)}</td></tr>
-            <tr><th>流程声明</th><td>${e(workflowDeclareText)}</td></tr>
-            <tr><th>流程异常</th><td>${e(workflowErrors)}</td></tr>
-          </tbody>
-        </table>
-      </details>`;
-    };
-    const renderReviewSummary = (): string => {
-      const groups: Array<[string, SkillDerivedStandard[]]> = [
-        ['待确认', sortedSoftStandards.filter((standard) => standard.status === 'pending_review')],
-        ['已确认', sortedSoftStandards.filter((standard) => standard.status === 'author_confirmed')],
-        ['已否决', sortedSoftStandards.filter((standard) => standard.status === 'rejected')],
-        ['已过期', sortedSoftStandards.filter((standard) => standard.status === 'stale')],
-      ];
-      const meta = record
-        ? `<div class="review-log-meta">
-            <div>生成模型：${e(record.model)}</div>
-            <div>生成时间：${e(record.generatedAt.slice(0, 19).replace('T', ' '))}</div>
-            <div>Prompt：${e(record.promptId)} / ${e(record.promptVersion)}</div>
-          </div>`
-        : '<p class="context-muted">还没有候选提取记录。</p>';
-      const grouped = groups.map(([label, standards]) => `<section class="review-log-group">
-        <h4>${e(label)} ${standards.length}</h4>
-        ${standards.length > 0
-          ? `<ul>${standards.map((standard) => `<li><span>${e(kindLabel(standard.standardKind))}</span><strong>${e(standard.title)}</strong></li>`).join('')}</ul>`
-          : '<p class="context-muted">暂无</p>'}
-      </section>`).join('');
-      return `${meta}${grouped}`;
-    };
-    const runtimeStatusIcon = (status: string): string =>
-      status === 'passed' ? '✅' : status === 'attention' ? '❌' : '⏳';
-    const runtimeStatusLabel = (status: string): string =>
-      status === 'passed' ? '有数据支撑' : status === 'attention' ? '有异常' : '待补充';
-    const runtimeStepList = (checks: typeof chain.runtime.hardRules, kind: 'workflow' | 'rule'): string => checks.length > 0
-      ? `<ol class="runtime-step-list">${checks.map((check, index) => `<li class="runtime-step runtime-${e(check.status)}">
-          <div class="runtime-step-head">
-            <span class="runtime-step-index">${kind === 'workflow' ? `第 ${index + 1} 步` : `规则 ${index + 1}`}</span>
-            <span class="runtime-step-name">${e(displayRuntimeCheckTitle(check.title, check.id, index))}</span>
-            <span class="runtime-step-state" title="${e(runtimeStatusLabel(check.status))}">${runtimeStatusIcon(check.status)}</span>
-          </div>
-          <details class="runtime-step-detail">
-            <summary>执行细节</summary>
-            <div class="runtime-step-detail-body">
-              ${shouldShowRuntimeCheckRawId(check.id) ? `<code>原始编号：${e(check.id)}</code>` : ''}
-              ${check.reason ? `<p>${e(check.reason)}</p>` : ''}
-              ${check.evidenceSnippets.length > 0 ? `<ul>${check.evidenceSnippets.map((snippet) => `<li>${e(snippet)}</li>`).join('')}</ul>` : ''}
-            </div>
-          </details>
-        </li>`).join('')}</ol>`
-      : '<p class="context-muted">暂无可检查项。</p>';
-    const contextChainNav = `<nav class="context-chain-nav" aria-label="切换查看小节"><a href="#cc-${e(skillName)}-1">① 原始能力定义</a><a href="#cc-${e(skillName)}-2">② 检测结果</a><a href="#cc-${e(skillName)}-3">③ 运行时</a><a href="#cc-${e(skillName)}-4">④ 人工复盘</a><span class="context-chain-nav-hint">可向右滚动查看</span></nav>`;
-    return `${contextChainNav}<div class="context-chain-grid">
-      <section class="context-chain-panel" id="cc-${e(skillName)}-1">
-        <h3>① 原始能力定义 / 候选标注</h3>
-        ${chain.definition.found
-          ? `<div class="context-meta">SKILL.md：<code>${e(chain.definition.path ?? '')}</code></div>
-             <pre class="skill-md-source">${renderAnnotatedSkillMd(chain.definition.content ?? '')}${chain.definition.truncated ? '\n\n... 已截断，仅展示前 12000 字符' : ''}</pre>
-             ${renderUnlocatedCandidates()}`
-          : '<p class="context-muted">未找到能力定义文件：当前日志里能识别到这个能力，但本机目录没有对应定义。属于"定义不可用"，不是"作者没声明规则"。</p>'}
-      </section>
-      <section class="context-chain-panel" id="cc-${e(skillName)}-2">
-        <h3>② 检测结果 / 标准化建议</h3>
-        ${missingNotices || '<p class="context-muted">已探测到标准化声明，下面先看流程执行线，再看规则清单。</p>'}
-        <h4>流程执行线</h4>
-        ${renderStructuredWorkflows()}
-        <h4>规则清单</h4>
-        ${renderStructuredHardRules()}
-        ${renderProbeLog()}
-      </section>
-      <section class="context-chain-panel" id="cc-${e(skillName)}-3">
-        <h3>③ 运行时</h3>
-        <p class="context-muted">${e(chain.runtime.message)}</p>
-        <table>
-          <tbody>
-            <tr><th>调用段</th><td>${chain.runtime.summary.invocationCount}</td></tr>
-            <tr><th>工具调用</th><td>${chain.runtime.summary.toolCallCount}</td></tr>
-            <tr><th>工具失败</th><td>${chain.runtime.summary.toolFailureCount}</td></tr>
-            <tr><th>证据符合</th><td>${chain.runtime.summary.passedCount}</td></tr>
-            <tr><th>需关注</th><td>${chain.runtime.summary.attentionCount}</td></tr>
-            <tr><th>无法自动判断</th><td>${chain.runtime.summary.manualReviewCount}</td></tr>
-          </tbody>
-        </table>
-        <h4>流程运行时证据</h4>
-        ${runtimeStepList(chain.runtime.workflowNodes, 'workflow')}
-        <h4>规则运行时证据</h4>
-        ${runtimeStepList(chain.runtime.hardRules, 'rule')}
-      </section>
-      <section class="context-chain-panel" id="cc-${e(skillName)}-4">
-        <h3>④ 人工复盘状态</h3>
-        ${renderReviewSummary()}
-      </section>
-    </div>`;
-  };
-  const renderConfidenceHeader = (padding = '8px 10px', border = '1px solid var(--border)'): string =>
-    `<th style="text-align:right;padding:${padding};border-bottom:${border}">
-      判断把握 / 归属把握
-      <span class="signal-help" tabindex="0" data-signal-title="判断把握 / 归属把握" data-signal-description="${e(confidenceHeaderHelp)}" aria-label="${e(confidenceHeaderHelp)}" style="display:inline-flex;align-items:center;justify-content:center;margin-left:4px;width:15px;height:15px;border:1px solid var(--border);border-radius:50%;font-size:10px;line-height:15px;color:var(--text-muted);cursor:help;vertical-align:middle">?</span>
-    </th>`;
-  const reviewPriorityMeta = (priority: ExperienceReviewPriority): { label: string; color: string; bg: string } => {
-    if (priority === 'review_first') return { label: '建议优先复盘', color: 'var(--red)', bg: 'rgba(220,38,38,.08)' };
-    if (priority === 'sample_review') return { label: '值得抽样复盘', color: 'var(--yellow)', bg: 'rgba(202,138,4,.10)' };
-    return { label: '常规抽样', color: 'var(--text-muted)', bg: 'var(--bg-muted)' };
-  };
-  const basisLabel = (code: ExperienceReviewBasisCode): string => {
-    const labels: Record<ExperienceReviewBasisCode, string> = {
-      has_high_observation: '有高优先级过程发现',
-      has_medium_observation: '有需要抽样确认的过程发现',
-      user_correction: '用户纠正过方向或结果',
-      user_interruption: '用户人工中断过当前执行',
-      session_interrupted: '会话出现异常中断信号',
-      negative_feedback: '用户出现负向反馈',
-      hard_rule_text_hit: '用户提出了必须/不要/禁止等硬性要求',
-      tool_failure: '运行中出现工具执行失败',
-      hedging_signal: '回答里出现不确定表达',
-      explicit_marker: '回答里明确标出未知/缺口',
-    };
-    return labels[code];
-  };
-  const renderPriorityBadge = (priority: ExperienceReviewPriority): string => {
-    const meta = reviewPriorityMeta(priority);
-    return `<span style="display:inline-flex;padding:3px 7px;border-radius:999px;background:${meta.bg};color:${meta.color};font-size:12px;font-weight:650">${e(meta.label)}</span>`;
-  };
-  const assistiveInferenceMeta = (code: ExperienceAssistiveInferenceCode): { label: string; description: string; className: string } => {
-    const values: Record<ExperienceAssistiveInferenceCode, { label: string; description: string; className: string }> = {
-      review_recommended: {
-        label: '辅助推断：需要人工复盘',
-        description: '确定性规则命中了用户纠正、人工中断、负向反馈、高优先级过程发现等信号。不是自动判定 skill 做错。',
-        className: 'assistive-attention',
-      },
-      sample_recommended: {
-        label: '辅助推断：建议抽样确认',
-        description: '确定性规则命中了工具执行失败、中等过程发现、不确定表达等信号，适合抽样看上下文。',
-        className: 'assistive-sample',
-      },
-      positive_signal_observed: {
-        label: '辅助推断：看到正向反馈',
-        description: '人工用户消息里出现正向反馈，且未命中更高优先级异常规则。',
-        className: 'assistive-positive',
-      },
-      user_switched_topic_neutral: {
-        label: '辅助推断：用户切换目标',
-        description: '人工用户消息里出现“换个方向/先不/不用这个”等目标切换表达。当前只记录目标中止或切走，不自动归因给 skill。',
-        className: 'assistive-unknown',
-      },
-      no_obvious_issue_from_rules: {
-        label: '辅助推断：未见明显异常',
-        description: '只基于当前固定规则，未发现需要优先复盘的异常信号；不代表最终质量结论。',
-        className: 'assistive-normal',
-      },
-      insufficient_human_context: {
-        label: '辅助推断：人工上下文不足',
-        description: '当前片段缺少足够人工用户消息，只能展示证据，不能推断用户预期是否满足。',
-        className: 'assistive-unknown',
-      },
-    };
-    return values[code];
-  };
-  const assistiveConfidenceLabel = (value: ExperienceAssistiveInference['confidence']): string => {
-    if (value === 'high') return '规则把握高';
-    if (value === 'medium') return '规则把握中';
-    return '规则把握低';
-  };
-  const assistiveCautionLabel = (code: ExperienceAssistiveInferenceCautionCode): string => {
-    const labels: Record<ExperienceAssistiveInferenceCautionCode, string> = {
-      no_llm_judge: '未使用 LLM 判断',
-      rule_only: '只用固定规则',
-      runtime_context_excluded: '运行时注入不计用户表达',
-      skill_context_excluded: 'Skill 文档注入不计入用户交互',
-      no_human_user_message: '缺少人工用户消息',
-      limited_timeline_window: '时间线为截断窗口',
-    };
-    return labels[code];
-  };
-  const renderAssistiveInference = (inference?: ExperienceAssistiveInference, compact = false, skillNameForAdvisory?: string): string => {
-    if (!inference) return '<span style="color:var(--text-muted)">暂无辅助推断</span>';
-    const meta = assistiveInferenceMeta(inference.code);
-    const basis = inference.basisRuleCodes.length > 0
-      ? inference.basisRuleCodes.map((code) => ruleFindingMeta(code).label).join('、')
-      : '未命中优先规则';
-    // 把 skill 级 advisory（缺 hardRules / 缺 workflows / SKILL.md 找不到）合并进 cautions：
-    // 这样 L1 表格里不点开 modal 也能从 hover 看到「标准缺失」这件事。
-    const skillAdvisoryCodes = skillNameForAdvisory ? collectSkillChainAdvisoryCodes(skillNameForAdvisory) : [];
-    const advisoryShortLabels = skillAdvisoryCodes.map((code) => getSkillChainAdvisory(code).shortLabel);
-    const cautionLabels = inference.cautionCodes.map(assistiveCautionLabel);
-    const cautions = [...advisoryShortLabels, ...cautionLabels].join('、') || '无';
-    const title = `${meta.description}\n依据：${basis}\n边界：${cautions}`;
-    if (compact) {
-      // L1 compact 模式不再展示「规则把握 high/medium/low」chip：
-      // 它与 assistive label 是同一份数据的复读，cautionCodes 在 hover title 里仍可见。
-      const advisoryChips = advisoryShortLabels.length > 0
-        ? `<div class="assistive-advisory-row">${advisoryShortLabels.map((label) => `<span class="assistive-advisory-chip" title="点开右侧「skill 体检」查看建议">⚠️ ${e(label)}</span>`).join('')}</div>`
-        : '';
-      return `<div class="assistive-box compact ${meta.className}" title="${e(title)}">
-        <div class="assistive-main">
-          <span>${e(meta.label)}</span>
-          <span class="signal-help assistive-help" tabindex="0" data-signal-title="${e(meta.label)}" data-signal-description="${e(title)}" aria-label="${e(title)}">?</span>
-        </div>
-        ${advisoryChips}
-      </div>`;
-    }
-    return `<div class="assistive-box ${meta.className}" title="${e(title)}">
-      <div class="assistive-main">
-        <span>${e(meta.label)}</span>
-        <strong>${e(assistiveConfidenceLabel(inference.confidence))}</strong>
-      </div>
-      <div class="assistive-desc">${e(meta.description)}</div>
-      <div class="assistive-sub">依据：${e(basis)}</div>
-      <div class="assistive-sub">边界：${e(cautions)}</div>
-    </div>`;
-  };
-  const reviewStateKey = (targetType: string, targetId: string): string => `${targetType}:${targetId}`;
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const renderReviewStateControls = (targetType: 'experience_session' | 'inbox_item' | 'skill' | 'goal_slice_correction', targetId: string): string => {
-    const key = reviewStateKey(targetType, targetId);
-    const entry = reviewState.entries[key];
-    const safeId = e(targetId);
-    const safeType = e(targetType);
-    const buttons = [
-      { verdict: 'reviewed', label: '已复盘', title: '标记为已经人工看过；再次点击可取消这个本地标记。' },
-      { verdict: 'real_issue', label: '确认问题', title: '标记为确认存在问题；可点击其他按钮直接改标记，再次点击可取消。' },
-      { verdict: 'not_issue', label: '不是问题', title: '标记为不是问题或无需处理；可点击其他按钮直接改标记，再次点击可取消。' },
-      { verdict: 'needs_more_context', label: '证据不足', title: '标记为证据不足，需要更多上下文；可点击其他按钮直接改标记，再次点击可取消。' },
-    ];
-    return `<div class="review-state-control" data-review-state-key="${e(key)}" data-review-state-current="${e(entry?.verdict ?? '')}">
-      <div class="review-state-actions">
-        ${buttons.map((button) => {
-          const active = entry?.verdict === button.verdict;
-          return `<button type="button" class="review-state-button ${active ? `is-active ${reviewVerdictClassName(button.verdict)}` : ''}" data-review-verdict="${button.verdict}" onclick="setObservationReviewState('${safeType}', '${safeId}', '${button.verdict}', this)" title="${e(button.title)}">${e(button.label)}</button>`;
-        }).join('')}
-      </div>
-    </div>`;
-  };
-  const reviewVerdictClassName = (verdict?: string): string => {
-    if (verdict === 'real_issue') return 'review-real-issue';
-    if (verdict === 'not_issue') return 'review-not-issue';
-    if (verdict === 'needs_more_context') return 'review-needs-context';
-    if (verdict === 'reviewed') return 'review-reviewed';
-    return '';
-  };
-  const softStandardStatusLabel = (status: SkillDerivedStandard['status']): string => {
-    if (status === 'author_confirmed') return '作者已确认';
-    if (status === 'rejected') return '已否决';
-    if (status === 'stale') return '已过期';
-    return '待作者确认';
-  };
-  const renderExperienceBasis = (codes: ExperienceReviewBasisCode[]): string => {
-    if (codes.length === 0) return '<span style="color:var(--text-muted)">没有触发复盘优先级规则；进入常规抽样池</span>';
-    return codes.map((code) => `<span style="display:inline-block;margin:0 4px 4px 0;padding:2px 6px;border-radius:999px;background:var(--bg-muted);color:var(--text-secondary);font-size:11px">${e(basisLabel(code))}</span>`).join('');
-  };
-  const ruleFindingMeta = (code: ExperienceRuleFindingCode): { label: string; description: string } => {
-    const values: Record<ExperienceRuleFindingCode, { label: string; description: string }> = {
-      high_observation_seen: { label: '高优先级过程发现', description: '存在 severity=high 的过程发现，建议优先复盘原始上下文。' },
-      medium_observation_seen: { label: '抽样过程发现', description: '存在 medium/low 过程发现，适合进入抽样复盘。' },
-      user_correction_seen: { label: '用户纠正', description: '人工用户消息命中明确纠偏表达。' },
-      user_interruption_seen: { label: '人工中断', description: 'trace 中出现用户主动中断事件。' },
-      session_interrupted_seen: { label: '会话异常中断', description: 'trace 中出现会话异常切换或 assistant turn failed 标记。' },
-      negative_feedback_seen: { label: '负向反馈', description: '人工用户消息命中明确负向表达。' },
-      positive_feedback_seen: { label: '正向反馈', description: '人工用户消息命中明确认可表达，仅作为正向证据展示。' },
-      user_goal_shift_seen: { label: '用户切换目标', description: '人工用户消息命中“换个方向/先不/不用这个”等目标切换表达；这通常表示目标切走，不自动归因给 skill。' },
-      hard_rule_seen: { label: '用户硬性要求', description: '人工用户消息命中“必须/不要/禁止”等临时硬性要求。Skill 自身的强约束请看定义链路里的规则检测结果。' },
-      tool_failure_seen: { label: '工具执行失败', description: 'tool_result 标记失败，表示执行过程中存在工具层失败。' },
-      hedging_seen: { label: '不确定表达', description: '过程发现里存在“不确定/可能/需要确认”等低置信文本信号。' },
-      explicit_marker_seen: { label: '显式缺口', description: '过程发现里存在“未知/知识缺口”等显式标记。' },
-      runtime_context_excluded: { label: '运行时注入已排除', description: 'SDK/工作台注入上下文已从用户交互指标里排除。' },
-      skill_context_excluded: { label: 'Skill 文档注入不计入用户交互', description: '系统会把 SKILL.md 内容注入给 agent 作为说明书；这不是用户本人说的话，所以不计入用户纠正、追问、情绪反馈或用户硬性要求。' },
-      no_priority_signal: { label: '未命中优先规则', description: '未命中需要优先复盘或抽样复盘的固定规则。' },
-    };
-    return values[code];
-  };
-  const ruleFindingLevelMeta = (level: ExperienceRuleFindingLevel): { label: string; className: string } => {
-    if (level === 'attention') return { label: '需关注', className: 'rule-attention' };
-    if (level === 'sample') return { label: '抽样看', className: 'rule-sample' };
-    return { label: '记录项', className: 'rule-normal' };
-  };
-  const renderRuleFindings = (findings: ExperienceRuleFinding[] = [], compact = false): string => {
-    if (findings.length === 0) return '<span style="color:var(--text-muted)">没有固定规则命中</span>';
-    return `<div class="${compact ? 'rule-finding-list compact' : 'rule-finding-list'}">${findings.map((finding) => {
-      const meta = ruleFindingMeta(finding.code);
-      const level = ruleFindingLevelMeta(finding.level);
-      const anchor = finding.evidenceRefs?.[0]?.messageIndex !== undefined ? ` · #${finding.evidenceRefs[0].messageIndex}` : '';
-      return `<span class="rule-finding ${level.className}" title="${e(meta.description)}">
-        <span class="rule-level">${e(level.label)}</span>
-        <span>${e(meta.label)}</span>
-        <strong>${finding.count}</strong>
-        <span class="rule-anchor">${e(anchor)}</span>
-      </span>`;
-    }).join('')}</div>`;
-    };
-    const fallbackEvidenceChain = (session: ExperienceSessionSummary): ExperienceEvidenceChain => {
-      const events = session.timelinePreview ?? [];
-      const userEvents = events.filter((event) => event.kind === 'user_message' && !isSyntheticUserMessageText(event.snippet ?? ''));
-      return {
-      userMessageCount: userEvents.length,
-      runtimeContextCount: events.filter((event) => event.kind === 'runtime_context').length,
-      skillContextCount: events.filter((event) => event.kind === 'skill_context').length,
-      assistantMessageCount: events.filter((event) => event.kind === 'assistant_message').length,
-      toolUseCount: events.filter((event) => event.kind === 'tool_use').length,
-      toolResultCount: events.filter((event) => event.kind === 'tool_result').length,
-      toolFailureResultCount: events.filter((event) => event.kind === 'tool_result' && event.isError === true).length,
-      observationCount: session.relatedObservationIds.length,
-    };
-  };
-  const renderEvidenceChain = (chain: ExperienceEvidenceChain): string => {
-    const item = (label: string, value: number, title: string): string =>
-      `<span class="evidence-chain-item" title="${e(title)}"><span>${e(label)}</span><strong>${value}</strong></span>`;
-    const anchor = (label: string, ref?: ExperienceEvidenceRef): string => {
-      if (!ref) return '';
-      const title = [ref.timestamp, ref.snippet].filter(Boolean).join(' · ');
-      const msg = ref.messageIndex !== undefined ? `#${ref.messageIndex}` : '有记录';
-      return `<button type="button" class="evidence-anchor" data-jump-message-index="${ref.messageIndex ?? ''}" data-jump-message-uuid="${e(ref.messageUuid ?? '')}" onclick="jumpToExperienceMessage(this)" title="${e(title || '切换并定位到对应消息')}">${e(label)} ${e(msg)}</button>`;
-    };
-    return `<div class="evidence-chain">
-      <div class="evidence-chain-row">
-        ${item('人工用户消息', chain.userMessageCount, '只包含人工用户原话，不包含 SDK 工作台注入或 SKILL.md 注入内容')}
-        ${item('运行时注入', chain.runtimeContextCount, 'SDK/工作台注入上下文，已从用户交互指标里排除')}
-        ${item('Skill 文档注入', chain.skillContextCount, 'SKILL.md 注入上下文，已从用户交互指标里排除')}
-        ${item('助手回复', chain.assistantMessageCount, 'assistant_message 上下文')}
-      </div>
-      <div class="evidence-chain-row">
-        ${item('工具调用', chain.toolUseCount, 'tool_use 事件数')}
-        ${item('工具执行结果', chain.toolResultCount, 'tool_result 事件数')}
-        ${item('失败结果', chain.toolFailureResultCount, 'tool_result 明确失败，或返回内容里带 status=error / error 字段')}
-        ${item('过程发现', chain.observationCount, '关联 observation item 数')}
-      </div>
-      <div class="evidence-anchor-row">
-        ${anchor('首条用户消息', chain.firstUserMessage)}
-        ${anchor('运行时注入', chain.firstRuntimeContext)}
-        ${anchor('Skill 文档', chain.firstSkillContext)}
-        ${anchor('首个工具调用', chain.firstToolUse)}
-        ${anchor('首个失败结果', chain.firstToolFailure)}
-        ${anchor('最后助手回复', chain.lastAssistantMessage)}
-      </div>
-    </div>`;
-  };
+  const observationMetricRenderers = createObservationMetricRenderers({ experience, reviewState });
+  const observationSignalRenderers = createObservationSignalRenderers(lang);
   const {
-    renderTimelinePair,
-  } = createObservationTimelineRenderers({
-    experience,
-    reviewState,
-    canonicalFeedbackSignalsForDisplay,
-    reviewStateKey,
-    shouldIncludeDownstreamFeedbackForDisplay,
+    evidenceConclusion,
+    renderEvidenceCell,
+    renderSignalLabel,
+    renderSourceBadge,
+    semanticEvidence,
+  } = observationSignalRenderers;
+  const observationSkillChainRenderers = createObservationSkillChainRenderers({
+    experienceToolCountsBySkill,
+    skillChains,
+    skillDerivedStandards,
+    metricRenderers: observationMetricRenderers,
   });
-  const renderReviewRows = (groupItems: ObservationInboxItem[], idPrefix: string): string => groupItems.map((item, index) => {
-    const evidence = semanticEvidence(item);
-    const detailsId = `${idPrefix}-detail-${index}`;
-    const searchText = [item.severity, item.sourceKind, item.skillName, item.signalType, item.signalSubtype, evidence, item.cwd, item.sourceTrace].join(' ').toLowerCase();
-    return `<tr data-observe-row="review" data-severity="${e(item.severity)}" data-search="${e(searchText)}" data-detail-id="${detailsId}">
-      <td style="padding:8px 10px">${renderSeverityBadge(item)}${renderSourceBadge(item)}</td>
-      <td style="padding:8px 10px">${renderSignalLabel(item)}</td>
-      <td class="num" style="padding:8px 10px;text-align:right">${renderOccurrences(item)}</td>
-      <td class="num" style="padding:8px 10px;text-align:right">${item.confidence.toFixed(2)} / ${item.attributionConfidence.toFixed(2)}</td>
-      <td style="padding:8px 10px;color:var(--text-muted);font-size:12px">${e(observedItemTimestamp(item, item.lastSeen))}</td>
-      <td style="padding:8px 10px">${renderEvidenceCell(item)}</td>
-      <td class="num" style="padding:8px 10px;text-align:right"><button type="button" onclick="toggleObservationDetail('${detailsId}', this)" style="font-size:12px;padding:4px 8px;border:1px solid var(--border);background:var(--bg);border-radius:4px;cursor:pointer">${lang === 'zh' ? '展开' : 'Details'}</button></td>
-    </tr>
-    <tr id="${detailsId}" data-observe-detail-for="${detailsId}" style="display:none;background:var(--bg-muted)">
-      <td colspan="7" style="padding:14px 18px;border-bottom:1px solid var(--border);text-align:left">
-        <div style="display:grid;grid-template-columns:minmax(0,0.75fr) minmax(0,1.55fr);gap:18px;align-items:start;max-width:100%;overflow:hidden">
-          <section style="text-align:left;font-size:11px;color:var(--text-muted);min-width:0">
-            <h3 style="font-size:12px;margin:0 0 8px;color:var(--text-secondary)">这条记录的来源和判断依据</h3>
-            ${renderField('id', item.id)}
-            ${renderField('sourceTrace', item.sourceTrace)}
-            ${renderField('sourceKind', item.sourceKind)}
-            ${renderField('sessionId', item.sessionId)}
-            ${renderField('cwd', item.cwd)}
-            ${renderArtifactVersion(item.artifactVersion)}
-            ${renderField('signalRule', signalRuleDescription(item))}
-            ${renderField('signalMeaning', `原始信号=${item.signalType}；OMK 判断出的失败原因=${item.signalSubtype}。例如 failed_search 表示工具搜索/读取失败，bash_probe/not_found/tool_limit 是 OMK 根据命令和输出内容进一步判断出来的原因。`)}
-            ${renderField('occurrencesMeaning', `出现次数=${item.occurrences}；按 skill/cwd/signal/subtype/query/path 聚合去重后的同类事件数量。`)}
-            ${renderField('reviewDecision', reviewSeverityMeta(item).decision)}
-            ${renderField('firstSeen', timestampedOccurrences(item) > 0 ? item.firstSeen : '未记录')}
-            ${renderField('lastSeen', timestampedOccurrences(item) > 0 ? item.lastSeen : '未记录')}
-            ${renderField('timestampCoverage', `${timestampedOccurrences(item)}/${item.occurrences}`)}
-            ${renderField('recentSessionIds', item.recentSessionIds.join(', '))}
-            <button type="button" onclick="openObservationTrace('${e(item.id)}', this)" style="margin-top:8px;font-size:12px;padding:5px 8px;border:1px solid var(--border);background:var(--bg);border-radius:4px;cursor:pointer">Open in trace</button>
-            <pre id="trace-${e(item.id)}" style="display:none;margin:8px 0 0;padding:9px;background:var(--bg-muted);border:1px solid var(--border);border-radius:6px;white-space:pre-wrap;word-break:break-word;font-size:11px;line-height:1.45;max-height:360px;overflow:auto;text-align:left"></pre>
-          </section>
-          <section style="text-align:left;min-width:0">
-            <h3 style="font-size:13px;margin:0 0 8px;color:var(--text-primary)">Evidence 明细</h3>
-            <div style="color:var(--text-muted);font-size:12px;margin-bottom:8px">这里展示这条聚合记录下面的原始明细。列表里每一条都是一次真实命中的证据；上方表格只展示第一条摘要。</div>
-            ${renderEvidenceDetailList(item)}
-            <details style="margin-top:10px">
-              <summary style="cursor:pointer;color:var(--text-muted);font-size:12px">查看原始 JSON</summary>
-              ${renderJson({ evidence: item.evidence, representativeEvidence: item.representativeEvidence })}
-            </details>
-          </section>
-        </div>
-      </td>
-    </tr>`;
-  }).join('');
+  const {
+    indicatorLabels,
+    indicatorHelps,
+    metric,
+    formatEntrypoint,
+    formatAttributionSource,
+    renderInvocationSummary,
+    compactRankedCountText,
+    renderOpenClawSourceMetadata,
+    renderSessionOpenClawSourceMetadata,
+    canonicalFeedbackSignalsForDisplay,
+    displayIndicatorsForSession,
+    sessionMetricSourceTitle,
+    displayBasisCodes,
+    displayPriority,
+    displayAssistiveInference,
+    displayRuleFindings,
+  } = observationMetricRenderers;
+  const {
+    renderSkillEvidenceSummary,
+    renderSkillChainSummary,
+    renderSkillChainButton,
+    renderRuntimeRuleFlow,
+    renderSkillChainTemplate,
+  } = observationSkillChainRenderers;
+  const {
+    fallbackEvidenceChain,
+    renderConfidenceHeader,
+    renderPriorityBadge,
+    renderAssistiveInference,
+    reviewStateKey,
+    renderExperienceBasis,
+    renderRuleFindings,
+    renderEvidenceChain,
+    renderReviewRows,
+    renderTimelinePair,
+  } = createObservationReviewRenderers({
+    experience,
+    lang,
+    observedItemTimestamp,
+    reviewState,
+    timestampedOccurrences,
+    metricRenderers: observationMetricRenderers,
+    signalRenderers: observationSignalRenderers,
+    skillChainRenderers: observationSkillChainRenderers,
+  });
   const skillGroups = Array.from(items.reduce((map, item) => {
     const existing = map.get(item.skillName) ?? [];
     existing.push(item);
@@ -2316,7 +545,6 @@ export function renderObservationInboxPage(model: ObservationInboxViewModel, lan
     </section>
   `).join('');
   const experienceGoalById = new Map((experience?.goalSlices ?? []).map((goal) => [goal.id, goal]));
-  const experienceSessionIdBySkillAndSession = new Map((experience?.sessions ?? []).map((session) => [`${session.skillName}\u0000${session.sessionId}`, session.id]));
   const metricGuideSections: Array<{ title: string; keys: IndicatorHelpKey[] }> = [
     { title: '用户交互', keys: ['userCorrection', 'userInterruption', 'userFollowUp', 'negativeFeedback', 'positiveFeedback', 'userGoalShift', 'hardRule'] },
     { title: 'Skill 执行', keys: ['toolCall', 'toolFailure', 'bash', 'read', 'grep'] },
