+++ test/architecture/import-boundaries.test.ts
@@ -173,8 +173,14 @@ const RULES: ForbiddenRule[] = [
       'studio/presentation/observation-inbox-renderer.ts::observability/inbox/view-model.ts',
       'studio/presentation/observation-inbox-renderer.ts::observability/inbox/feedback-projection.ts',
       'studio/presentation/observation-inbox/helpers.ts::observability/inbox/feedback-projection.ts',
+      'studio/presentation/observation-inbox/metric-renderer.ts::observability/inbox/view-model.ts',
+      'studio/presentation/observation-inbox/metric-renderer.ts::observability/inbox/feedback-projection.ts',
+      'studio/presentation/observation-inbox/review-renderer.ts::observability/inbox/view-model.ts',
+      'studio/presentation/observation-inbox/review-renderer.ts::observability/inbox/feedback-projection.ts',
       'studio/presentation/observation-inbox/reviewer-report.ts::observability/inbox/view-model.ts',
       'studio/presentation/observation-inbox/reviewer-report.ts::observability/inbox/feedback-projection.ts',
+      'studio/presentation/observation-inbox/signal-renderer.ts::observability/inbox/view-model.ts',
+      'studio/presentation/observation-inbox/skill-chain-renderer.ts::observability/inbox/view-model.ts',
       'studio/presentation/observation-inbox/timeline.ts::observability/inbox/view-model.ts',
       'studio/presentation/observation-inbox/timeline.ts::observability/inbox/feedback-projection.ts',
       'studio/presentation/skill-health-renderer.ts::observability/skill-health/analyzer.ts',
